% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_MT_omrader.R, R/copy_MT_omrader.R,
%   R/read_MT_omrader.R
\name{add_MT_omrader}
\alias{add_MT_omrader}
\alias{copy_MT_omrader}
\alias{read_MT_omrader}
\title{Manage translation from komnr to MT-omrader}
\usage{
add_MT_omrader(
  data,
  year = as.numeric(format(Sys.Date(), "\%Y")),
  fag = NULL,
  translation_table = komnr_2_MT_omrader,
  code_column = c("komnr"),
  new_column = c("MT_avdelingnr", "MT_avdeling", "MT_seksjonnr", "MT_seksjon"),
  position = "right",
  overwrite = FALSE,
  shortname = FALSE
)

copy_MT_omrader(
  filename = list("komnr_2_MT_avdeling.csv", "MT_omrader.csv", "komnr_2_MT_enhet.csv"),
  from_path = base::paste0(set_dir_NVI("GrunndataLand"), "FormaterteData/"),
  to_path = NULL
)

read_MT_omrader(
  filename = list("komnr_2_MT_avdeling.csv", "MT_omrader.csv", "komnr_2_MT_enhet.csv"),
  from_path = file.path(set_dir_NVI("GrunndataLand", slash = FALSE), "FormaterteData"),
  ...
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data with a column with kommunenummer (komnr).}

\item{year}{[\code{integer(1)}]\cr
year for which the organisation of MT should be selected. Defaults to current
    year.}

\item{fag}{[\code{character(1)}]\cr
fag is the production or topic in the organisation of MT. This is used for
    year >= 2025. Defaults to \code{NULL}.}

\item{translation_table}{[\code{data.frame}]\cr
The translation table for translating komnr to MT-areas. Defaults
    to \code{komnr_2_MT_omrader}.}

\item{code_column}{[\code{character(1)}]\cr
The column with the code value. Valid values are one
    of c("komnr", "MT_avdelingnr", "MT_regionnr", "MT_divisjonnr"). If the column in data
    has another name, it can be input as a named vector, see examples.
    Defaults to "komnr".}

\item{new_column}{[\code{character}]\cr
The name(s) of the new column(s) that should be added to the data,
    see examples. Defaults to c("MT_avdelingnr", "MT_avdeling",
    "MT_seksjonnr", "MT_seksjon").}

\item{position}{[\code{character(1)}]\cr
The position for the new columns, can be one of c("first", 
    "left", "right", "last", "keep"). The input can be 
    abbreviated, but must be unique, i.e. c("f", "le", 
    "r", "la", "k"). Defaults to "right".}

\item{overwrite}{[\code{logical(1)}]\cr
When the new column(s) already exist, 
    the content in the existing column(s) is replaced by new data if 
    \code{overwrite = TRUE}. If the new column(s) already exist and 
    \code{overwrite = FALSE}, an error is issued. Defaults to \code{FALSE}.}

\item{shortname}{[\code{logical(1)}]\cr
If \code{TRUE}, a short version of the MT unit name is generated, see details.
    Defaults to \code{FALSE}.}

\item{filename}{[\code{list}]\cr
File names of the source files for the translation table. Defaults to
    list("komnr_2_MT_avdeling.csv", "MT_omrader.csv", "komnr_2_MT_enhet.csv").}

\item{from_path}{[\code{character(1)}]\cr
Path to the source file(s) used to generate the translation table. Defaults 
    to their standard directory at NVI's internal network.}

\item{to_path}{[\code{character(1)}]\cr
Path to where the source file(s) used to generate the translation table should
    be copied.}

\item{\dots}{Other arguments to be passed to
\ifelse{html}{\code{\link[utils:read.csv2]{utils::read.csv2}}}{\code{utils::read.csv2}}.}
}
\value{
\code{add_MT_omrader} A data frame where the number and name of the
    MT-areas have been added in the column to the right of the column with
    the komnr.

\code{read_MT_omrader} A data frame with the table for translating from komnr
    to MT-omrader as read from the source csv file. If not changing standard
    input to the function, the standard files at NVI's internal network is read.

\code{copy_MT_omrader} Copies the csv-files "komnr_2_MT_avdeling.csv",
    "MT_omrader.csv", and "komnr_2_MT_enhet.csv" to another directory. If the
    target files already exists the source files are only copied if they are
    newer than the target files.
}
\description{
Function to add columns with the various MT-omrader. In addition,
   there are functions to read and copy the translation tables.
}
\details{
\code{add_MT_omrader} can be used to translate the komnr into
    MT-omrader. The organisation of MT in areas was revised in 2025. If the
    input year >= 2025, the komnr will be translated to the structure from 2025:
    MT_divisjonnr, MT_divisjon, MT_avdelingnr,MT_avdeling, MT_seksjonnr,
    MT_seksjon. If the input year is between 2004 and 2024, the komnr will be
    translated to the structure from 2004 to 2024: MT_avdelingnr, MT_avdeling,
    MT_regionnr and MT_region.

The function can also be used to translate MT_avdelingnr into
    MT_avdeling, MT_regionnr and MT_region or to translate MT_regionnr
    into MT_region.

When the \code{code_column} in the data frame is
    not named one of c("komnr", "MT_avdelingnr", "MT_regionnr") the
    \code{code_column} can be input as a named vector. Likewise, if
    the new columns should be given other names than c("MT_avdelingnr",
    "MT_avdeling", "MT_regionnr", "MT_region"), the \code{new_column}
    can be input as a named vector, see examples.

\code{add_MT_omrader} uses a pre made translation tables
    ("komnr_2_MT_avdeling.csv" and "komnr_2_MT_enhet.csv"). These data need
    to be loaded by \code{read_MT_omrader} before running \code{add_MT_omrader},
    see example. "komnr_2_MT_avdeling.csv" is made based on information in
    PJS adresseregister and "komnr_2_MT_enhet.csv" has been given by MT in
    2025. The translation table is updated when we know there is a need.

\code{position} is used to give the place if the new columns in the
    data frame. For \code{position = "right"} the new variables are
    placed to the right of the code_variable. Likewise, for
    \code{position = "left"} the new variables are placed to the left of
    the code_variable. If \code{position = "first"} or
    \code{position = "last"} the new columns are placed first or last,
    respectively, in the data.frame. A special case occurs for
    \code{position = "keep"} which only has meaning when the new column
    has the same name as an existing column and overwrite = TRUE. In these
    cases, the existing column will be overwritten with new data and have
    the same position.

If \code{shortname = TRUE} a short version of the MT unit name is generated.
    I.e. "Divisjon", "Avdeling" and "Seksjon" is removed from the unit name.
    \code{shortname = TRUE} has only meaning for the MT units from 2025 and
    onwards.

\code{read_MT_omrader} reads the files "komnr_2_MT_avdeling.csv",
    "MT_omrader.csv", and "komnr_2_MT_enhet.csv" into a data frame, usually
    named komnr_2_MT_omrader. This file is used by \code{add_MT_omrader}. If
    no options to the
    function is given, the function will read the latest updated files
    from NVI's internal network. If changing the \code{from_path}, the
    function can be used to read the translation file from other
    directories. This can be useful if having a script that don't have
    access to NVI's internal network.

\code{copy_MT_omrader} Copies the csv-files "komnr_2_MT_avdeling.csv",
    "MT_omrader.csv", and "komnr_2_MT_enhet.csv" to another directory.
    Thereby, these files are available for \code{read_MT_omrader} if they
    should be read from another directory.
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
komnr_2_MT_omrader <- read_MT_omrader()

# Copy the csv-files used to generate the translation table from the standard location to
# the subdirectory Data below the working directory
copy_MT_omrader(to_path = "./Data/")

# Reading from the subdirectory Data below the working directory
komnr_2_MT_omrader <- read_MT_omrader(from_path = "./Data/")

# Add new columns with MT_avdelingnr, MT_avdeling, MT_regionnr, and MT_region based on komnr
# Remember to load "komnr_2_MT_omrader" by "read_MT_omrader()" before running "add_MT_omrader",
# see above.
newdata <- add_MT_omrader(olddata,
                       translation_table = list(komnr_2_MT_omrader),
                       code_column = "komnr",
                       new_column = c("MT_avdelingnr", "MT_avdeling", "MT_regionnr", "MT_region"))

# Add new columns with MT_avdelingnr and MT_avdeling based on komnr. The colname of the column
# with komnr is komnr and the new columns are renamed to MT_avdnr and MT_avd.
# Remember to load "komnr_2_MT_omrader" by "read_MT_omrader()" before running "add_MT_omrader",
# see above.
newdata <- add_MT_omrader(olddata,
                       translation_table = list(komnr_2_MT_omrader),
                       code_column = c("komnr" = "komnr"),
                       new_column = c("MT_avdnr" = "MT_avdelingnr", "MT_avd" = "MT_avdeling"))

# Add new columns with MT_region based on MT_regionnr. MT_region is renamed to MT_regionnavn
# Remember to load "komnr_2_MT_omrader" by "read_MT_omrader()" before running "add_MT_omrader",
# see above.
newdata <- add_MT_omrader(olddata,
                       translation_table = list(komnr_2_MT_omrader),
                       code_column = "MT_region",
                       new_column = c("MT_regionnavn" = "MT_region"))
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
