% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_produsent-deprecated.R
\name{add_produsent}
\alias{add_produsent}
\title{add_produsent is Deprecated}
\usage{
add_produsent(
  data,
  translation_table,
  code_column,
  new_column,
  position = "right",
  overwrite = FALSE
)
}
\arguments{
\item{data}{Data frame with data with a column with a prodnr8}

\item{translation_table}{Data frame with the table for translating from prodnr8 to gjeldende_prodnr8.}

\item{code_column}{The column with the coded value. Valid values are one of c("prodnr8"). If the column in
data has another name, it can be input as a named vector, see examples.}

\item{new_column}{The new columns that should be included into the dataframe. The new columns can be up to
c("gjeldende_prodnr8") depending on \code{code_column}. If the new columns in the result dataframe
should have other names, \code{new_column} can be input as a named vector, see examples.}

\item{position}{position for the new columns, can be one of c("first", "left", "right", "last", "keep")}

\item{overwrite}{When the new column(s) already exist, the content in the existing column(s) is replaced by new data if overwrite = TRUE.
If the new columns already exists and overwrite = FALSE, then an error is issued.}
}
\description{
add_produsent was deprecated 2022-05-02. as other properties 
    than 'gjeldende_prodnr8' could not be included without breaking backward 
    compatibility. Use \code{add_produsent_properties} instead and ensure 
    to set the parameter \code{impute_old_when_missing = TRUE} when translating 
    from "prodnr8" to "gjeldende_prodnr8" and set the parameter 
    \code{impute_old_when_missing = FALSE} when translating from "prodnr8" to 
    other properties.
}
\details{
The old help pages can be found at \code{help("add_produsent-deprecated")}. 
    Information on deprecated function can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
