% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_arguments.R
\name{assert_add_function}
\alias{assert_add_function}
\title{Collection of assertions for add_functions}
\usage{
assert_add_function(
  data,
  translation_table,
  code_column,
  new_column,
  position,
  overwrite
)
}
\arguments{
\item{data}{Argument to the add-function to be asserted.}

\item{translation_table}{Argument to the add-function to be asserted.}

\item{code_column}{Argument to the add-function to be asserted.}

\item{new_column}{Argument to the add-function to be asserted.}

\item{position}{Argument to the add-function to be asserted.}

\item{overwrite}{Argument to the add-function to be asserted.}
}
\value{
\code{TRUE} if none of the assertions failed. If any of the assertions
    failed, one or more error messages are returned.
}
\description{
Collection of assertions used in standard add_functions.
}
\details{
All add functions except one, have the same arguments and the
    assertion can be standardized.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
