% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_arguments.R
\name{assert_add_functions}
\alias{assert_add_functions}
\title{Collection of assertions for add_functions}
\usage{
assert_add_functions(
  data,
  translation_table,
  code_column,
  new_column,
  overwrite,
  add
)
}
\arguments{
\item{data}{Argument to the add-function to be asserted.}

\item{translation_table}{Argument to the add-function to be asserted.}

\item{code_column}{Argument to the add-function to be asserted.}

\item{new_column}{Argument to the add-function to be asserted.}

\item{overwrite}{Argument to the add-function to be asserted.}

\item{add}{assertCollection.}
}
\value{
An assertCollection that have been updated with the results
    of assertions for \code{data}, \code{translation_table},
    \code{code_column}, \code{new_column} and \code{overwrite}.
}
\description{
Collection of assertions used in standard add_functions.
}
\details{
All add functions except one, have the same arguments and the
    assertion can be standardized.

The assertCollection must be set up separately and is used as input argument, see example.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
