% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_arguments.R
\name{assert_read_functions}
\alias{assert_read_functions}
\title{Collection of assertions for read_functions}
\usage{
assert_read_functions(filename, from_path, add = checks)
}
\arguments{
\item{filename}{Argument to the read-function to be asserted.}

\item{from_path}{Argument to the read-function to be asserted. Ending
"\\" and "/" should have been removed before the assertion is performed.}

\item{checks}{AssertCollection, defaults to checks.}
}
\value{
\code{TRUE} if none of the assertions failed. If any of the assertions
    failed, one or more error messages are returned. An AssertCollection that have been updated with the result of the checks.
}
\description{
Collection of assertions used in standard read_functions.
}
\details{
All read functions have the same arguments and the
    assertion can be standardized. The assertion of \code{filname} is
    constructed to handle both character and list input.

    The assertion is based on removing ending "\\" and "/" from
    \code{from_path} before the assertion is performed.
The function does not set up and report the assertCollection.
  This need to be done separately, see example.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
