% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_query_hensikt.R
\name{build_query_hensikt}
\alias{build_query_hensikt}
\title{Builds query for selecting data for hensikt from PJS}
\usage{
build_query_hensikt(year, hensikt, db = "PJS")
}
\arguments{
\item{year}{The year that should be selected as integer. Can be given as one year, the first and last year or a range of years.}

\item{hensikt}{Vector with one or more hensiktkoder. Sub-hensikter are not included and must be explicitly mentioned.}

\item{db}{The database for which the query is built. Currently only the value "PJS" is accepted.}
}
\value{
A list with select-statements for v2_sak_m_res and v_sakskonklusjon, respectively. The statements should
   be included in a \code{RODBC::sqlQuery}.
}
\description{
Builds the query for selecting all data for one or more hensikt within one year from PJS. The query is
   written in T-SQL as used by MS-SQL.
}
\details{
The function builds the SQL syntax to select all PJS-journals concerning the hensiktkoder from PJS.
}
\examples{
# SQL-select query for Pancreatic disease (PD)
build_query_hensikt(year = 2020,
                    hensikt = c("0200102"))
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
