% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_query_one_disease-deprecated.R
\name{build_query_one_disease}
\alias{build_query_one_disease}
\title{build_query_one_disease is Deprecated}
\usage{
build_query_one_disease(
  year,
  analytt,
  hensikt = NULL,
  metode = NULL,
  db = "PJS"
)
}
\arguments{
\item{year}{[\code{numeric}]\cr
One year or a vector giving the first and last years that should 
be selected.}

\item{analytt}{[\code{character}]\cr
Analyttkoder that should be selected. If sub-analytter should be included,
end the code with \%.}

\item{hensikt}{[\code{character}]\cr
Specific hensiktkoder. If sub-hensikter should be included,
end the code with \%. Defaults to \code{NULL}.}

\item{metode}{[\code{character}]\cr
Specific metodekoder. Defaults to \code{NULL}.}

\item{db}{[\code{character(1)}]\cr
The database for which the query is built. Defaults to "PJS" that currently
is the only valid value.}
}
\description{
\code{build_query_one_disease} was deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::build_query_one_disease} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
The old help pages can be found at \code{help("build_query_one_disease-deprecated")}.
    Information on deprecated functions can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
