% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sql_modules.R
\name{build_sql_modules}
\alias{build_sql_modules}
\alias{build_sql_select_year}
\alias{build_sql_select_code}
\title{Builds sql modules to be included in select statements for PJS}
\usage{
build_sql_select_year(year, varname, db = "PJS")

build_sql_select_code(values, varname, db = "PJS")
}
\arguments{
\item{year}{One year or a vector with the first and last year that should
be selected as integer vector.}

\item{varname}{The PJS variable name of the variable in PJS from which the
coded values should be selected.}

\item{db}{The database for which the query is built. Currently only the value "PJS" is accepted.}

\item{values}{The value of the codes that should be selected given as character.
If sub-codes should be included, add "\%" after the code, see example.}
}
\value{
SQL-code to be included when building select-statements for PJS.
}
\description{
Builds sql modules to be included in select statements for PJS
when building queries for selecting data. The functions takes the values
for which observations should be selected as input and builds the
sql syntax.
}
\details{
\code{build_sql_select_year} builds the SQL syntax to select observations
from one or more consecutive years from PJS. The input can be given as
one year, the first and last year or a range of years. If a range is given,
this will be interpreted as first and last years and all years in between
will be included.

\code{build_sql_select_code} builds the SQL syntax to select observations
with the given code values from one variabel in PJS with hierarchical codes.
When the code value including sub codes  should be selected, add "\%" to the
code, see example.

Be aware that these functions only builds an sql building block to be
included into a select statement. It will not build a complete select
statement. These functions are mainly intended for internal use and
are called from \code{\link{build_query_hensikt}} and \code{\link{build_query_one_disease}}.
If generating own select statements, these can be used to facilitate the
coding. The building blocks can be combined with "AND" and "OR" and
brackets to get the intended select statement.
}
\examples{
# SQL-select module for selecting year from PJS
build_sql_select_year(year = 2020, varname = "aar")

build_sql_select_year(year = c(2019, 2021), varname = "aar")

build_sql_select_year(year = c(2019:2021), varname = "aar")

# SQL-select module for selecting hensiktkode from PJS
build_sql_select_code(values = "0100101", varname = "hensiktkode", db = "PJS")

build_sql_select_code(values = "0100101\%", varname = "hensiktkode", db = "PJS")

build_sql_select_code(values = c("0100101", "0100101007", "0100102\%", "0100202\%"),
                      varname = "hensiktkode",
                      db = "PJS")

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
