% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sql_modules-deprecated.R
\name{build_sql_select_year}
\alias{build_sql_select_year}
\title{build_sql_select_year is Deprecated}
\usage{
build_sql_select_year(year, varname, db = "PJS")
}
\arguments{
\item{year}{[\code{numeric}]\cr
One year or a vector giving the first and last years that should 
be selected.}

\item{varname}{[\code{character(1)}]\cr
The PJS variable name of the variable in PJS from which the
coded values should be selected.}

\item{db}{[\code{character(1)}]\cr
The database for which the query is built. Defaults to "PJS" that currently
is the only valid value.}
}
\description{
\code{build_sql_select_year} was deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::build_sql_select_year} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
The old help pages can be found at \code{help("build_sql_modules-deprecated")}.
    Information on deprecated function can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
