% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_PJS_levels-deprecated.R
\name{choose_PJS_levels}
\alias{choose_PJS_levels}
\title{choose_PJS_levels is Deprecated}
\usage{
choose_PJS_levels(
  data,
  levels,
  keep_col = NULL,
  remove_col = NULL,
  unique_rows = TRUE
)
}
\arguments{
\item{data}{Data frame with data from PJS}

\item{levels}{PJS-levels from which data should be chosen. Valid values are c("sak", "prove", "delprove", "undersokelse", "resultat",
"konklusjon", "subundersokelse", "subresultat").}

\item{keep_col}{Column names of columns that should be included in addition to the columns defined by levels.}

\item{remove_col}{Column names of columns that should be removed even if being at the defined levels.}

\item{unique_rows}{If \code{TRUE} (default), only unique rows are included in the data frame.}
}
\description{
\code{choose_PJS_levels} was deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::choose_PJS_levels} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
The old help pages can be found at \code{help("choose_PJS_levels-deprecated")}.
    Information on deprecated functions can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
