% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cut_slash}
\alias{cut_slash}
\title{Cut away ending slash from string}
\usage{
cut_slash(x)
}
\arguments{
\item{x}{Object with character strings.}
}
\value{
Object without ending slash in character strings.
}
\description{
Removes ending slash or backslash from string.
    This is used to clean pathnames so that elements in a path
    can be combined using \code{file.path} in stead of \code{\link{paste0}}.
}
\examples{
# Remove from string
cut_slash("C:/temp/")
cut_slash("C:\\\\temp\\\\")
cut_slash(c("C:/temp/", "C:\\\\temp\\\\"))
cut_slash(list("C:/temp/", "C:\\\\temp\\\\"))
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
