% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_from_PJSdata-defunct.R
\name{exclude_from_PJSdata-defunct}
\alias{exclude_from_PJSdata-defunct}
\title{exclude rows from PJS-data}
\arguments{
\item{PJSdata}{Data frame with data extracted from PJS.}

\item{abroad}{If equal "exclude", samples from abroad are excluded. Allowed
values are c("exclude", "include").}

\item{quality}{If equal "exclude", samples registered as quality assurance
and ring trials are excluded. Allowed values are c("exclude", "include").}
}
\value{
data frame without excluded PJS-data.
}
\description{
Performs common subsetting of PJS-data by excluding rows
}
\details{
Performs common cleaning of PJSdata by removing samples that usually
    should not be included when analyzing PJSdata. The cleaning is dependent
    on having the following columns eier_lokalitettype, eierlokalitetnr and
    hensiktkode.

    \code{abroad = "exclude"} will exclude samples that have eier_lokalitet
    of type "land" and eier_lokalitetnr being different from NO. Samples
    registered on other types than LAND are not excluded.

    \code{quality = "exclude"} will exclude all samples registered s quality
    assurance and ring trials, i.e. hensiktkode starting with "09".
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
