% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_file_near_date.R
\name{find_file_near_date}
\alias{find_file_near_date}
\title{Finds a file based on date in file name}
\usage{
find_file_near_date(
  from_path,
  partial_filename = NULL,
  extension = NULL,
  date_position = "last",
  wanted_date = Sys.Date(),
  wanted_year = NULL,
  wanted_month = NULL,
  nearest_wanted = "before"
)
}
\arguments{
\item{from_path}{[\code{character(1)}]\cr
Full path for the directory with the files. Defaults to \code{NULL}.}

\item{partial_filename}{[\code{character}]\cr
One or more patterns within the file name that will be used to identify a
   subset of files in the directory that can be selected. Defaults to
   \code{NULL}.}

\item{extension}{[\code{character}(1)]\cr
The file extension of the files that can be selected. Defaults to \code{NULL}.}

\item{date_position}{[\code{character}(1)]\cr
The position of the date in the file name. Must be one of \code{c("last", "per")}.
    Defaults to \code{"last"}.}

\item{wanted_date}{[\code{character}(1)]\cr
The date for which the register date should be nearest. Defaults to
    \code{Sys.Date()}.}

\item{wanted_year}{[\code{character}(1)] | [\code{numeric}(1)]\cr
The year from which the register should be read, see details. Defaults to
    \code{NULL}.}

\item{wanted_month}{[\code{character}]\cr
The month for which the register should be read, see details. Defaults to
    \code{NULL}.}

\item{nearest_wanted}{[\code{character(1)}]\cr
Whether the nearest register date before the wanted date or the nearest date
    after the wanted date should be selected, see details. Must be one of
    \code{c("before", "after")}. Defaults to \code{"before"}.}
}
\value{
A data frame with the file name of the file that fulfills the
    selection criteria.
}
\description{
Identifies the file name of a file based on date information in
    the file name. This is usually used to identify the file name of a file
    with data that has a register date nearest the wanted date. The register
    date must be included in the file name and written as _per_DATE, see
    details.
}
\details{
Reads the file names of files in a given directory and selects the
    file nearest a wanted date based on date information in the file name.
    The register date is used to identify the file name of the data that either
    has the last register date before the wanted date (the default) or the
    first register date after the wanted date.
The function assumes that the register date for the data is written in the
    file name either as a "DATE" placed last in the file name before the extension
    or as "_per_DATE" that can be written anywhere in the file name. DATE should
    be given either as a date (formatted as \%Y\%m\%d), a year_month
    (formatted as \%Y\%m) or a year (formatted as \%Y).
If one want the last register date before the wanted date and the
    "per_DATE" is given as year, the file will be identified as before the
    wanted date for any wanted date within the register year. Likewise, If
    one want the last register date after the wanted date and the
    "per_DATE" is given as year, the file will be identified as after the
    wanted date for any wanted date within the register year. I.e. for the
    file name "data_per_2025.csv" this file name will be found for the dates
    c("2025-01-01", "2025-12-31") and all between for both
    nearest_to_per_date = "before" and nearest_to_per_date = "after". This
    assumes that there are no other files for that year with a more precise
    register date.
If one want the last register date before the wanted date and the
    "per_DATE" is given as year_month, the file will be identified as before
    the wanted date for any wanted date within the register month. Likewise,
    If one want the last register date after the wanted date and the
    "per_DATE" is given as year_month, the file will be identified as after the
    wanted date for any wanted date within the register month. I.e. for the
    file name "data_per_202503.csv" this file name will be found for the dates
    c("2025-03-01", "2025-03-31") and all between for both
    nearest_to_per_date = "before" and nearest_to_per_date = "after". This
    assumes that there are no other files for that year_month with a more
    precise register date.
}
\examples{
\dontrun{
# Identify the filename of the last Purkering register before November 2023
filelist <- find_file_near_date(
  from_path = file.path(NVIdb::set_dir_NVI("Ekst", slash = FALSE),
  "Purkeringer", "FormaterteData"),
  partial_filename = c("Purkering"),
  extension = "csv",
  date_position = "last",
  wanted_date = NULL,
  wanted_year = 2023,
  wanted_month = "11",
  nearest_wanted = "before")
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
