% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_columns.R
\name{find_standard_values_for_property}
\alias{find_standard_values_for_property}
\title{Find standard column values for given property}
\usage{
find_standard_values_for_property(
  column_names,
  dbsource,
  org_values_in_column = NULL,
  new_values_in_column = NULL,
  standard
)
}
\arguments{
\item{column_names}{[\code{character}]\cr
The column names of the data source. Defaults to \code{colnames(data)}.}

\item{dbsource}{[\code{character(1)}]\cr
The database that is the source of data. Should be the name of
    the data source as registered in column_standards table. Defaults
    to \code{dbsource}.}

\item{org_values_in_column}{[\code{character(1)}]\cr
The column name in the standards table in which the column names for which
    one will find the property values. Must be one
    of c("colname_db", "colname"). Defaults to \code{NULL}.}

\item{new_values_in_column}{[\code{character(1)}]\cr
The column name in the standards table in which one will find the property
    values. Must be one of c("colname", "colclass", "collabel", "colwidths_Excel",
    "colorder"). Defaults to \code{NULL}.}

\item{standard}{[\code{data.frame}]\cr
Table with column_standards. Defaults to column_standards.}
}
\value{
A named vector where previously unnamed elements have been named with
    the element value as name.
}
\description{
Ensures that all elements in a vector are named.
}
\details{
Sub-function called by standardize columns to standardise code.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
