% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_file_near_date.R
\name{last_date_in_month}
\alias{last_date_in_month}
\title{Identifies the last date in a month}
\usage{
last_date_in_month(date = NULL, year = NULL, month = NULL, output = "date")
}
\arguments{
\item{date}{[\code{date}(1)]\cr
The date for which the last date should be identified. Defaults to
\code{NULL}.}

\item{year}{[\code{numeric(1)} | \code{character(1)}]\cr
The year of the month for which the last date should be identified.
Defaults to \code{NULL}.}

\item{month}{[\code{character(1)}]\cr
The month within the year for which the last date should be
identified. Defaults to \code{NULL}.}

\item{output}{[\code{character(1)}]\cr
Should the output be the last date or the last day of the month. Must be
one of c("date", "day"). Defaults to \code{"date"}.}
}
\value{
The last date or the last day of the month depending on value for
    output.
}
\description{
Identifies the last date or the last day for a given month
    within a year.
}
\details{
The function is restricted to the period from year 1700 to year
    2099. One must either input a date or the month and year.
One can chose if the output should be the last date within the month or the
    last day number within the month.
}
\examples{
\dontrun{
last_date_in_month(date = as.Date("2024-01-24"), output = "date")
last_date_in_month(as.Date("2024-01-24"), output = "day")
last_date_in_month(year = "2024", month = "02", output = "date")
last_date_in_month(year = "2024", month = "02", output = "day")
last_date_in_month(as.Date("2025-02-01"), output = "date")
last_date_in_month(as.Date("2025-02-28"), output = "day")
last_date_in_month(year = "2025", month = "08", output = "date")
last_date_in_month(year = "2025", month = "08", output = "day")
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
