% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_EOS-deprecated.R
\name{login_EOS}
\alias{login_EOS}
\alias{login_by_input_EOS}
\alias{login_by_credentials_EOS}
\title{The login_EOS-functions are deprecated}
\usage{
login_EOS(dbinterface = NULL, ...)

login_by_input_EOS(dbinterface = NULL, ...)

login_by_credentials_EOS(dbinterface = NULL, ...)
}
\arguments{
\item{dbinterface}{The R-package that is used for interface towards the data
base.}

\item{\dots}{Other arguments to be passed from the wrappers to
login_by_credentials or login_by_input}
}
\description{
\code{login_EOS}, \code{login_by_input_EOS}, and
    \code{login_by_credentials_EOS}, were deprecated from v0.##.# released 2024-##-##.
    These wrapper functions were rarely used and therefore considered unnecessary.
}
\details{
The old help pages can be found at \code{help("login_EOS-deprecated")}.
    Information on deprecated function can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
