% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_PJS-deprecated.R
\name{login_PJS}
\alias{login_PJS}
\alias{login_by_input_PJS}
\alias{login_by_credentials_PJS}
\title{The login_PJS-functions are deprecated}
\usage{
login_PJS(dbinterface = NULL, ...)

login_by_input_PJS(dbinterface = NULL, ...)

login_by_credentials_PJS(dbinterface = NULL, ...)
}
\arguments{
\item{dbinterface}{The R-package that is used for interface towards the data
base.}

\item{\dots}{Other arguments to be passed from the wrappers to
login_by_credentials or login_by_input.}
}
\description{
\code{login_PJS}, \code{login_by_input_PJS}, and
    \code{login_by_credentials_PJS}, were deprecated from v0.##.# released 2024-##-##.
    The wrapper functions \code{login_PJS} and \code{login_by_input_PJS} were
    rarely used and therefore considered unnecessary. Although
    \code{login_by_credentials_PJS} have been used extensively, it can easily
    be replaced by \code{login_by_credentials("PJS")}. Therefore, it has been
    considered unnecessary and have been deprecated.
}
\details{
The old help pages can be found at \code{help("login_PJS-deprecated")}.
    Information on deprecated function can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
