% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{order_fylke}
\alias{order_fylke}
\title{Data: Fylke sorted in logical order for reporting}
\format{
An object of class \code{character} of length 33.
}
\source{
"Norwegian Statistics"
}
\usage{
order_fylke
}
\description{
A vector of fylke names sorted in logical order for reporting. The vector
    includes all fylker and areas used in statistic as well as historical fylker back to
    1970.
}
\details{
The sort order are from south - east, south and northwards. The sorting is an
    adaption of the order implemented for fylker from 1960 to 2018 as reflected in the
    fylke numbers for this period.

    The vector is based on the different edition of fylke from Norwegian Statistics.
    However, the fylke name does not include the sami part of the name.

    If the order of fylker should be changed, one should edit the variable sortering_fylke
    in the source file. The code for generating the vector is written in
    "./data-raw/generate_ordered_fylke.R".
}
\keyword{datasets}
