% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PJS_code_description_colname}
\alias{PJS_code_description_colname}
\title{Data: PJS_code_description_colname, standard column names for description texts for selected code variables in PJS.}
\format{
A data frame with 3 variables:
\describe{
  \item{code_colname}{column name for selected code variables in PJs and that have been standardized using \code{NVIdb::standardize_columns}}
  \item{type}{the type of PJS variable as used by \code{NVIdb::add_PJS_code_description} to translate PJS-codes to description text}
  \item{new_column}{The new standard column names for the corresponding code column name in PJS}
}
}
\source{
\code{./data-raw/generate_PJS_code_description_colname.R} in package \code{NVIdb}
}
\usage{
PJS_code_description_colname
}
\description{
A data frame with the variable names (column names) in PJS and their corresponding PJS-level. The column names
   are the standardized column names, i.e. after running \code{NVIdb::standardize_columns}.
   The raw data can be edited in the \code{./data-raw/generate_PJS_code_description_colname.R}. The \code{PJS_code_description_colname}
   is used by \code{NVIdb::add_PJS_code_description} when using the options \code{PJS_variable_type = "auto"} and/or
   \code{new_column = "auto"}.
}
\keyword{datasets}
