% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_avlsgris.R
\name{read_avlsgris}
\alias{read_avlsgris}
\title{Read Register with avlsgrisbesetninger}
\usage{
read_avlsgris(
  from_path = file.path(set_dir_NVI("EksterneDatakilder", slash = FALSE), "Avlsgris",
    "FormaterteData"),
  year = "last",
  month = NULL,
  ...
)
}
\arguments{
\item{from_path}{[\code{character(1)}]\cr
Path for the produksjonstilskuddsregister. Defaults to the standard
directory at the NVI network.}

\item{year}{[\code{character(1)}] | [\code{numeric(1)}]\cr
The year from which the register should be read. Options is "last", or
a year. Defaults to "last".}

\item{month}{[\code{character(1)}]\cr
The month for which the register should be read. Defaults to \code{NULL}.}

\item{\dots}{Other arguments to be passed to
\ifelse{html}{\code{\link[data.table:fread]{data.table::fread}}}{\code{data.table::fread}}.}
}
\value{
A \code{data.frame} with the avlsgris-register.
}
\description{
Functions to read versions of the
    avlsgris-register.
}
\details{
The avlsgris-register includes information on the avlsgris herds.
    The register is updated from the industry at least once a year. This
    function automatically selects the last updated
    version of the register.

    \code{read_avlsris} reads the avlsgris-register into a
    data frame. The function gives options to select year and month. If there
    are no available version from the selected month, the last available
    version before the chosen month, will be selected. The
    standard settings will read in the files from NVI's internal network and
    select the latest updated file.
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
avlsgris <- read_avlsgris()

# Reading from standard directory at NVI's network and
#     selecting a specific version of the register
avlsgris2021 <- read_avlsgris(year = 2021, Pkode_month = "03")
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
