% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_csv_file}
\alias{read_csv_file}
\title{Read csv-file}
\usage{
read_csv_file(filename, from_path, options = NULL, data.table = FALSE, ...)
}
\arguments{
\item{filename}{[\code{character(1)}]\cr
File name of the source file that should be read.}

\item{from_path}{[\code{character(1)}]\cr
Path for the source file.}

\item{data.table}{[\code{logical(1)}]\cr
Should the result be of class \code{data.table} (\code{TRUE}) or
    \code{data.frame} (\code{FALSE}). Defaults to \code{FALSE}.}

\item{\dots}{Other arguments to be passed to \code{data.table::fread}.}

\item{columnclasses}{Predefine format (numeric or character) of the variables.}

\item{fileencoding}{usually UTF-8.}
}
\value{
A data frame with the data from the source file.
}
\description{
Reads files with data using \code{data.table::fread}.
}
\details{
Internal function used in read-functions. Uses \code{data.table::fread}
   that is much faster than \code{utils::read.table}. The default output is
   a \code{data.frame}, but this can be changed to a \code{data.table} by
   setting the argument \code{data.table = FALSE}.
}
\examples{
\dontrun{
read_csv_file(filename, from_path, columnclasses, fileencoding)
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
