% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eos_data-defunct.R
\name{read_eos_data-defunct}
\alias{read_eos_data-defunct}
\title{Read EOS data from RaData}
\arguments{
\item{from_path}{[\code{character(1)}]\cr
Path for raw data from eos_data.}

\item{eos_table}{[\code{character(1)}]\cr
The name of the table with eos raw data.}

\item{year}{[\code{character} | \code{numeric}]\cr
The years to be included in the result. Can be both numeric
or character. Defaults to \code{NULL}, i.e. no selection.}

\item{colClasses}{[\code{character}]\cr
The class of the columns, as in
\ifelse{html}{\code{\link[utils:read.table]{utils::read.table}}}{\code{utils::read.table}}.
Defaults to \code{"character"}.}

\item{encoding}{[\code{character(1)}]\cr
The encoding, one of c("UTF-8", "latin1"). Defaults to \code{"UTF-8"}.}

\item{\dots}{Other arguments to be passed to
\ifelse{html}{\code{\link[data.table:fread]{data.table::fread}}}{\code{data.table::fread}}.}
}
\value{
A data frame with data from EOS.
}
\description{
Reads EOS data from RaData. Includes historical data if these exists.
    It is possible to limit the data to one or more years.
}
\details{
read_eos_data uses
    \ifelse{html}{\code{\link[data.table:fread]{data.table::fread}}}{\code{data.table::fread}}
    to read the data with the settings \code{showProgress = FALSE} and
    \code{data.table = FALSE}. Other arguments can be passed to
    \ifelse{html}{\code{\link[data.table:fread]{data.table::fread}}}{\code{data.table::fread}}
    if necessary.

The eos_table name is the same name as the name as in the EOS data base.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
