% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eos_data.R
\name{read_eos_data}
\alias{read_eos_data}
\title{Read EOS data from RaData}
\usage{
read_eos_data(
  eos_table,
  from_path = paste0(set_dir_NVI("EOS"), "RaData"),
  year = NULL,
  colClasses = "character",
  encoding = "UTF-8",
  ...
)
}
\arguments{
\item{eos_table}{The name of the table with eos raw data}

\item{from_path}{Path for raw data from eos_data}

\item{year}{The years to be included in the result. Can be both numeric
or character. Defaults to \code{NULL}, i.e. no selection.}

\item{colClasses}{The class of the columns, as in utils::read.table, Defaults to 
\code{"character"}.}

\item{encoding}{The encoding. Defaults to \code{"UTF-8"}.}

\item{\dots}{Other arguments to be passed to \code{data.table::fread}.}
}
\value{
A data frame with data from EOS.
}
\description{
Reads EOS data from RaData. Includes historical data if these exists.
    It is possible to limit the data to one or more years.
}
\details{
read_eos_data uses \code{data.table::fread} to read the data with the 
    settings \code{showProgress = FALSE} and \code{data.table = FALSE}. Other 
    arguments can be passed to \code{data.table::fread} if necessary.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
