% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_leveransereg.R
\name{read_leveransereg}
\alias{read_leveransereg}
\title{Read Leveranseregisteret for slakt}
\usage{
read_leveransereg(
  filename,
  from_path = paste0(set_dir_NVI("LevReg"), "FormaterteData/"),
  ...
)
}
\arguments{
\item{filename}{[\code{character(1)}]\cr
Name of the file with Leveranseregisteret.}

\item{from_path}{[\code{character(1)}]\cr
Path for the Leveranseregisteret. Defaults to the standard
directory at the NVI network.}

\item{\dots}{Other arguments to be passed to
\ifelse{html}{\code{\link[data.table:fread]{data.table::fread}}}{\code{data.table::fread}}.}
}
\value{
\code{read_LevReg} A data frame with Leveranseregisteret as in selected csv-file.
}
\description{
Functions to read Leveranseregisteret for slakt.
}
\details{
The Leveranseregisteret for slakt includes information on carcasses
    delivered to slaughter. The register include identity of the farmer,
    slaughterhouse, date of slaughter, animal species, category (age group
    and sex), and weight. For poultry the individual animal is not reported,
    but number of slaughtered poultry per categories, slaughterhouse and date.

\code{read_leveransereg} reads the Leveranseregisteret for slakt into a
    data frame. The standard settings will read in the files from NVI's
    internal network. If changing the \code{from_path}, the function can be
    used to read Leveranseregisteret from other directories. This can be
    useful if having a stand alone app with no connection the NVI's internal
    network. In other cases, it should be avoided.
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
LevReg2019 <- read_leveransereg(filename = "LevReg_201901_201912.csv")
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
