% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_purkering.R
\name{read_purkering}
\alias{read_purkering}
\title{Read the register of purkeringer}
\usage{
read_purkering(
  from_path = file.path(set_dir_NVI("EksterneDatakilder", slash = FALSE), "Purkeringer",
    "FormaterteData"),
  wanted_date = Sys.Date(),
  wanted_year = NULL,
  wanted_month = NULL,
  nearest_wanted = "before",
  ...
)
}
\arguments{
\item{from_path}{[\code{character(1)}]\cr
Full path for the directory with the purkering-register files. Defaults to
    standard directory at NVI's home network.}

\item{wanted_date}{[\code{character}(1)]\cr
The date for which the register date should be nearest. Defaults to
    \code{Sys.Date()}.}

\item{wanted_year}{[\code{character}(1)] | [\code{numeric}(1)]\cr
The year from which the register should be read, see details. Defaults to
    \code{NULL}.}

\item{wanted_month}{[\code{character}]\cr
The month for which the register should be read, see details. Defaults to
    \code{NULL}.}

\item{nearest_wanted}{[\code{character(1)}]\cr
Whether the nearest register date before the wanted date or the nearest date
    after the wanted date should be selected, see details. Must be one of
    \code{c("before", "after")}. Defaults to \code{"before"}.}

\item{\dots}{Other arguments to be passed to
\ifelse{html}{\code{\link[utils:read.csv2]{utils::read.csv2}}}{\code{utils::read.csv2}}.}
}
\value{
A \code{data.frame} with the purkering-register.
}
\description{
Function to read the register of purkeringer, i.e.
    the central herd and the satellite herds in the sow pools.
}
\details{
The purkering-register includes information on the herds in the sow
    pools. The register is updated from the industry at least once a year.

\code{read_purkering} reads the purkering-register into a \code{data.frame}. The
    default is to read the last updated version of the register. You may change
    this by the input to wanted date or wanted year and month to read the
    register nearest (either before or after) the wanted date. See help for
    \code{\link{find_file_near_date}} for a full description of the use of
    wanted date or wanted year and month.
}
\examples{
\dontrun{
# Reading the last version from the standard directory at NVI's network
purkering <- read_purkering()

# Reading from standard directory at NVI's network and
#     selecting a specific version of the register
purkering2021 <- read_purkering(wanted_year = 2021, wanted_month = "12")
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
