% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varekode.R
\name{read_varekode}
\alias{read_varekode}
\title{Manage translation table for varekoder til leveransregisteret}
\usage{
read_varekode(
  filename = "varekoder.csv",
  from_path = paste0(set_dir_NVI("LevReg")),
  year = NULL,
  data_source = "formatted"
)
}
\arguments{
\item{filename}{Name of the translation table, defaults to "varekoder.csv".
The input is only used when \code{data_source = "formatted"}.}

\item{from_path}{Path for the translation table for varekoder.}

\item{year}{Year(s) for fetching the varekoderegister.}

\item{data_source}{Reads formatted data or raw data. deafult is formatted.}
}
\value{
\code{read_varekoder} A data frame with the translation table for
    varekoder to descriptive text and metadata.
}
\description{
Read the translation table for varekoder til leveransregisteret.
}
\details{
The translation table for varekoder comprises the variables:
    the leveranseaar, varekode, vare (descriptive text), dyreslag,
    vareart, dyrekategori, and varekategorikode. The register covers 2016
    and later.

    \code{read_varekoder} with the argument \code{type = "formatted"}
    reads the formatted "varekoder.csv" into a data frame. The standard
    settings will read the file from NVI's internal network. If changing
    the from_path, the function can be used to read the translation file
    from other directories. This can be useful if having a stand alone
    app with no connection the NVI's internal network. In other cases,
    it should be avoided.

    \code{read_varekoder} with the argument \code{type = "raw"} reads
    the raw data as supplied from Landbruksdirektoratet into a data frame.
    Thereafter, these can be used to generate the formatted version.
    The standard settings will read the file from NVI's internal network
    and changing the path should be avoided.
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
varekoder <- read_varekode()
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
