% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_objects.R
\name{remove_obects}
\alias{remove_obects}
\title{Remove objects if they exist in the environment}
\usage{
remove_obects(object, envir = environment())
}
\arguments{
\item{object}{[\code{character}]\cr
Objects that should be removed.}

\item{envir}{[\code{environment}]\cr
The environment from which the objects should be removed. Defaults to
    environment(), i.e. the environment from which the
    function was called.}
}
\value{
none, objects are removed from the environment.
}
\description{
Convenience wrapper around \code{rm}. The function accepts a vector
    with named objects and checks if the objects exist before trying removing
    them.
}
\examples{
\dontrun{
library(NVIdb)
xx <- 12345
remove_objects("xx")

xy <- 123
xz <- 456
remove_objects(c("xy", "xz", "zz"))
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
