% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_PJSdata-deprecated.R
\name{retrieve_PJSdata}
\alias{retrieve_PJSdata}
\title{retrieve_PJS is Deprecated}
\usage{
retrieve_PJSdata(
  year = NULL,
  selection_parameters = NULL,
  FUN = NULL,
  select_statement = NULL,
  ...
)
}
\arguments{
\item{year}{[\code{numeric}]\cr
One year or a vector giving the first and last years that should be selected.
    Defaults to \code{NULL}.}

\item{selection_parameters}{[\code{character(1)}]\cr
Either the path and file name for an R script that can be sourced and that
    sets the selection parameters or a named list with the selection parameters
    (i.e. of the same format as the output of
    \ifelse{html}{\code{\link{set_disease_parameters}}}{\code{set_disease_parameters}}).
    Defaults to \code{NULL}.}

\item{FUN}{\code{deprecated}\cr
\code{FUN} should instead be included as input to \code{selection_parameters}.
    Defaults to \code{NULL}.}

\item{select_statement}{\code{deprecated}\cr
\code{select_statement} should instead be included as input to
    \code{selection_parameters}. Defaults to \code{NULL}.}

\item{\dots}{Other arguments to be passed to the underlying functions:
\ifelse{html}{\code{\link{login_PJS}}}{\code{login_PJS}}
and
\ifelse{html}{\code{\link{exclude_from_PJSdata}}}{\code{exclude_from_PJSdata}}.}
}
\description{
\code{retrieve_PJS} was deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::retrieve_PJS} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
The old help pages can be found at \code{help("retrieve_PJSdata-deprecated")}.
    Information on deprecated functions can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
