% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_avlsgris.R
\name{select_file_for_date}
\alias{select_file_for_date}
\title{Selects files based on date in file name}
\usage{
select_file_for_date(
  from_path,
  filename_text,
  file_extension,
  year,
  month = "12",
  day = NULL,
  match_date = "last_before",
  extracted_date = NULL
)
}
\arguments{
\item{from_path}{[\code{character(1)}]\cr
Path for the files.}

\item{filename_text}{[\code{character}]\cr}

\item{file_extension}{[\code{character}]\cr}

\item{year}{[\code{character}] | [\code{numeric}]\cr
The year(s) from which the register should be read. Options is "last", or a
    vector with one or more years.}

\item{month}{[\code{character}]\cr
The month for which the register should be read, see details. Defaults to
\code{NULL}.}

\item{day}{[\code{character}]\cr}

\item{match_date}{[\code{character(1)}]\cr}

\item{extracted_date}{[\code{character(1)}]\cr}
}
\value{
A data frame with file names of the files that should be selected.
}
\description{
List file names that are selected based on the date information
    in the file name.
}
\details{
Reads the file names of files with the register date version in the
    file name. The routine assumes that the first number in the file name
    that can be a year, is the year (\%Y), year_month (\%Y\%m) or
    year_month_day (\%Y\%m\%d) for the register date. The register date
    is extracted from the file name and the file name of correct source
    file(s) are thereafter selected.

    The function is called from \code{read_avlsgris}.
}
\examples{
\dontrun{
# Making the filelist for read_Prodtilskudd or copy_Prodtilskudd
filelist <- select_prodtilskudd_files(from_path = from_path,
                                      Pkode_year = Pkode_year,
                                      Pkode_month = Pkode_month)
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
