% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_PJSdata_for_value.R
\name{select_PJSdata_for_value}
\alias{select_PJSdata_for_value}
\title{Selects a subset of PJSdata based on a vector of values}
\usage{
select_PJSdata_for_value(
  data,
  code_column,
  value_2_check,
  keep_selected = TRUE
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
PJS data from which a subset should be selected.}

\item{code_column}{[\code{character}]\cr
Vector with the column names for the variables that is used in the selection.}

\item{value_2_check}{[\code{character}]\cr
Vector with the values that should be selected, see details and examples.}

\item{keep_selected}{[\code{logical(1)}]\cr
If \code{TRUE}, the selected rows are included, if \code{FALSE}, the selected columns
are excluded. Defaults to \code{TRUE}.}
}
\value{
A \code{data.frame}.
}
\description{
Selects a subset of PJSdata based on a vector of values.
}
\details{
Selects a subset of PJSdata based on a vector of values.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
