% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_prodtilskudd_files}
\alias{select_prodtilskudd_files}
\title{List selected files from Søknad om register for produksjonstilskudd}
\usage{
select_prodtilskudd_files(from_path, Pkode_year, Pkode_month)
}
\arguments{
\item{from_path}{Path for the source translation table for PJS-codes}

\item{Pkode_year}{The year(s) from which the register should be read. Options is "last", or a vector with one or more years.}

\item{Pkode_month}{the month for which the register should be read. The options are c("05", "10", "both", "last") for Pkode_year = 2017
and c("03", "10", "both", "last") for Pkode_year >= 2018.}
}
\value{
A data frame with filenames of the files with the selected extracts of Prodtilskudd.
}
\description{
List selected files with extracts from Søknad om register for produksjonstilskudd.
}
\details{
Reads the filenames of files with extracts from Søknad om register for produksjonstilskudd into a data frame.
    The function gives options to select year and month and path for the files. The function is called from read_Prodtilskudd
    and copy_Prodtilskudd.
}
\examples{
\dontrun{
# Making the filelist for read_Prodtilskudd or copy_Prodtilskudd
filelist <- select_prodtilskudd_files(from_path = from_path,
                                      Pkode_year = Pkode_year,
                                      Pkode_month = Pkode_month)
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
