% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials_EOS-deprecated.R
\name{set_credentials_EOS}
\alias{set_credentials_EOS}
\title{set_credentials_EOS is deprecated}
\usage{
set_credentials_EOS()
}
\description{
\code{set_credentials_EOS} was deprecated from v0.##.# released 2024-##-##.
     The function was rarely used and is an unnecessary wrapper around
     \code{\link{set_credentials}}.
}
\details{
The old help pages can be found at \code{help("set_credentials_EOS-deprecated")}.
    Information on deprecated functions can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
