% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dir_NVI.R
\name{set_dir_NVI}
\alias{set_dir_NVI}
\title{Set directories for data sources at NVI}
\usage{
set_dir_NVI(datasource)
}
\arguments{
\item{datasource}{The data source that one want to access. The input can be 
abbreviated and case is ignored. To identify short names for the available 
directories, use \code{set_dir_NVI(datasource = "?")}.}
}
\value{
The full path for the directory at NVI's network. The path ends with "/".
}
\description{
Set the directories for various data sources at NVI's network.
}
\details{
The Norwegian Veterinary Institute has standard data sources at fixed 
    directories. The function returns the standard directory for the given 
    data source. Thereby hard coding of the paths may be avoided.
}
\examples{
\dontrun{
# Set path_ProdTilskudd to path for Prodtilskudd at the NVI network
prodtilskudd_path <- set_dir_NVI(datasource = "ProdTilskudd")
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
