% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_disease_selection_parameters.R
\name{set_disease_selection_parameters}
\alias{set_disease_selection_parameters}
\title{Sets disease selection parameters}
\usage{
set_disease_selection_parameters(
  year = NULL,
  hensikt2select = NULL,
  utbrudd2select = NULL,
  metode2select = NULL,
  analytt2select = NULL,
  file = NULL
)
}
\arguments{
\item{year}{One year or a vector with years giving the first and last years
that should be selected as integer.}

\item{hensikt2select}{Vector with specific hensikter. If sub-hensikter should
be included, end the code with \%. Can be \code{NULL}.}

\item{utbrudd2select}{String with an utbrudd ID. Can be \code{NULL}.}

\item{metode2select}{Vector with specific metoder. Can be \code{NULL}.}

\item{analytt2select}{Vector with  one or more analyttkode given as a character.
If sub-analytter should be included, end the code with \%. Can be \code{NULL}.}

\item{file}{An R script that can be sourced and that sets the parameters
hensikt2select, utbrudd2select, metode2select, analytt2select. Can be
\code{NULL}.}
}
\value{
A list with select-statement fom v2_sak_m_res and v_sakskonklusjon to
    be included in a \code{RODBC::sqlQuery}.
}
\description{
Sets the disease selection parameters in a list object. The list
    follows a standardised and named format and can be used as input to
    build_query_one_disease or build_query_hensikt.
}
\details{
Saker in PJS that concern one infection / disease can be characterised
    by the analytt (at konklusjon and/or resultat level), specific hensikter,
    utbrudds_ID and/or specific metoder. These can be used to select saker in
    PJS and/or structurise and simplify the output from PJS.

    One or more specific hensikter may be input to the selection statement.
    With specific hensikt is meant a hensikt that will imply
    that the sample will be examined for the infectious agent or disease.
    One or more specific metoder may be input to the selection statement.
    With specific metode is meant a metode that implies an
    examination that will give one of the input analytter as a result.

    The selection parameters can be input the function as such. Alternatively,
    a source file for the input parameters hensikt2select, utbrudd2select,
    metode2select and analytt2select may be given. This may be handy if the
    selection will be performed many times. It also gives the possibility of
    using a for loop for selecting and analysing data for several similar
    diseases at one time.

    The ouput is a named list with the selection parameters. This list can be
    used as input build_query_one_disease or build_query_hensikt.
}
\examples{
# Selection parameters for Pancreatic disease (PD)
selection_parameters <- set_disease_selection_parameters(
  year = 2020,
  analytt2select = c("01220104\%", "1502010235"),
  hensikt2select = c("0100108018", "0100109003", "0100111003", "0800109"),
  metode2select = c("070070", "070231", "010057", "060265")
  )
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
