% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_columns.R
\name{standardize_columns}
\alias{standardize_columns}
\title{Standardize columns for scripts and reports}
\usage{
standardize_columns(
  data,
  dbsource = deparse(substitute(data)),
  standards = NULL,
  property,
  language = "no",
  exclude = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame or if \code{property = "colclasses"} the path and filname of the  csv-file used as data source}

\item{dbsource}{database source of data. Set to data if not specifically specified. Needed if translation to column names is dependent on data source}

\item{standards}{to input alternative standard tables to column_standards}

\item{property}{Property of the column that should be standardized, currently c("colnames", "colclasses", "collabels", "colwidths_Excel", "colorder").}

\item{language}{Language for labels. Valid input are c("no", "en")}

\item{exclude}{Used in combination with \code{property = "colorder"}. \code{exclude = TRUE} excludes all columns with no predefinedcolorder.}

\item{\dots}{Other arguments to be passed read.csv2 when \code{property = "colclasses"}.}
}
\value{
\code{property = "colnames"}. A data frame with standard column names.

    \code{property = "colclasses"}. a named vector of column classes to be used as input to functions for reading csv-files.

    \code{property = "collabels"}. a vector with labels for the columns in the data frame.

    \code{property = "colwidths_Excel"}. a vector with column widths for Excel. To be used as input parameter to \code{openxlsx::.colwidth()}.

    \code{property = "colorder"}. A data frame with column names in predefined order. If exclude = TRUEonly columns withh a defined order is included
}
\description{
Standardizes column names, labels, column width for variables in external databases.
}
\details{
Experimental, the standardization table is under development. This version only works when being connected to the NVI network.

    Variables in internal and external data sources uses different variable names for the same content.
    \code{Standarddize_columns} standardizes column names for use in scripts. It will be further developed to standardize  column labels
    and column widths for both Excel and DT. Furthermore, input values for the parameter \code{colClasses = } for \code{read.csv2} can be
    generated.

    \code{property = "colnames"} will replace the column names in a data frame with standardized column names.
    All standard column names is snake_case. If no standard name is defined for a variable name, the variable
    name is translated to snake_case and the national characters \code{c("æ", "ø", "å")} are translated to \code{c("ae", "oe", "aa")}.

    \code{property = "colclasses"} will generate a named vector with the column classes for variables that may not be read correct when importing
    data from a csv-file. This applies for example to numbers with leading zero that must be imported as character. This vector can be used as a
    parameter for \code{colClasses = }. 
    
    The default fileEncoding is assumed to be "UTF-8". If another encoding one must give an additional argument like \code{fileEncoding = "latin"}.

    \code{property = "collabels"} will generate a vector with column labels that can be used to replace the column names in the header of the data
    table. The column names are not changed automatiacally but can be changed by using a colname statement (see help). If no standard column label
    is defined, the column name as Sentence case is used as column label. If English names are used and no English column label exists, the Norwegian
    column label is used instead.

    \code{property = "colwidths_Excel"} will generate a vector with column widths for Excel. To be used as input parameter to \code{openxlsx::.colwidth()}.
    If no standard column width is defined, the Excel standard width of 10.78 is used. Be aware that the generation of column widths are based on the
    column names. Do not change the column names to labels before the column widths are generated.

    \code{property = "colorder"} will generate a data frame with  the column names in a predefined order. The column names should first have been standardized.
    No standard order will be given unless the dbsource is defined in the column_standards table. If \code{exclude = FALSE} (the standard) the columns with no
    predefined order will be moved to the last columns in the same order as they appeared in the original data frame. If \code{exclude = TRUE} all columns with
    no predefined order is excluded from the data frame. This option is mainly intended for well defined and worked through routines like making selections lists
    for the Food Safety Authority. Do not use \code{exclude = TRUE} unless you are certain that all columns that should be included are defined in the
    column_standards table for this dbsource. If uncertain, you may first try with \code{exclude = FALSE} and thereafter compare with \code{exclude = TRUE} to
    check if you loose important information.
}
\examples{
\dontrun{
# Generate data frame to be standardized
df <- cbind("\u00C5r" = 2020, "Hensiktkode" = "01001", komnr = "5001")
colnames(df)

# Standardize column names
df <- standardize_columns(data = df, property = "colnames")
colnames(df)

# Generate vector with standard labels
labels <- standardize_columns(data = df, property = "collabels")
# use the labels as column names
colnames(df) <- labels

# Generate vector with standard column widths for Excel
colwidths <- standardize_columns(data = df, property = "colwidths_Excel")
colwidths

}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
