% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_eos_data.R
\name{standardize_eos_data}
\alias{standardize_eos_data}
\title{Standardising EOS-data}
\usage{
standardize_eos_data(
  data,
  dbsource = deparse(substitute(data)),
  standards = NULL,
  breed_to_species = TRUE,
  adjust_n_examined = TRUE,
  ...
)
}
\arguments{
\item{data}{Data frame with data retrieved from EOS.}

\item{dbsource}{If specified, this will be used for fetching standard column
names by \code{\link{standardize_columns}}. Defaults to the name of data.}

\item{standards}{data.frame The translation table to standard column names. 
Defaults to \code{NULL}.}

\item{breed_to_species}{[logical(1)] If \code{TRUE}, breed is translated back 
to species. Defaults to \code{TRUE)}.}

\item{adjust_n_examined}{[logical(1)] If \code{TRUE}, the number of examined 
samples is adjusted so it is maximum number of received samples. Defaults 
to \code{TRUE)}.}

\item{\dots}{Other arguments to be passed to \code{standardize_columns}.}
}
\value{
data frame with standardized EOS-data.
}
\description{
Standardising EOS-data. This standardising should always be performed.
    Otherwise the summary numbers can be wrong.
}
\details{
The function performs the following standardising of data extracted from EOS:
\itemize{
  \item The column names are standardised using \code{\link{standardize_columns}}.
  \item Numeric variables are transformed to numbers.
  \item Date variables are transformed to date format.
  \item Character variables are trimmed for leading and trailing spaces.
  \item Double registrations of a Sak due to the municipality being divided 
        between two Food Safety Authority offices are merged into one and for 
        these the information on Food Safety Authority office is removed.
  \item Breed is transformed to species. 
  }
}
\examples{
\dontrun{
# Standardizing sak_m_res
proveresultat_bse <- standardize_eos_data(data = proveresultat_bse)
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no

Johan Åkerstedt Johan.Akerstedt@vetinst.no
}
