% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_PJSdata.R
\name{standardize_PJSdata}
\alias{standardize_PJSdata}
\title{Standardizing PJS-data}
\usage{
standardize_PJSdata(PJSdata, dbsource = "v2_sak_m_res")
}
\arguments{
\item{PJSdata}{Data frame with data extracted from PJS.}

\item{dbsource}{If specified, this will be used for standardize columns when selecting standard column names}
}
\value{
data frame with standardized PJS-data.
}
\description{
Standardizing PJS-data that always should be performed.
}
\details{
The function performs the following standardizing of data extracted from PJS:
    - The unnecessary columns konkl_provenr and vet_distriktnr are removed,
    - The column names are standardized using standardize_columns,
    - Numeric variables are transformed to numbers,
    - Date variables are transformed to date format,
    - Character variables are trimmed for leading and trailing spaces,
    - The variables saksnr and, if possible, fagnr are generated,
    - Test data, i.e. saker with ansvarlig_seksjon in c("14", "99") or are deleted.
}
\examples{
\dontrun{
# Standardizing sak_m_res
 sak_m_res <- standardize_PJSdata(PJSdata = sak_m_res)
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no

Johan Åkerstedt Johan.Akerstedt@vetinst.no
}
