% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_code_combinations-deprecated.R
\name{transform_code_combinations}
\alias{transform_code_combinations}
\title{transform_code_combinations is Deprecated}
\usage{
transform_code_combinations(
  data,
  from_values,
  to_values,
  impute_when_missing_from = NULL
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data with code values that should be transformed.}

\item{from_values}{[\code{list}]\cr
List with named vector(s) of code values that should transformed, see details and examples.}

\item{to_values}{[\code{list}]\cr
List with named vector(s) of code values that should be the results of the transformation,
    see details and examples.}

\item{impute_when_missing_from}{[\code{character}]\cr
Column names for the code variables from which code values should be copied if no
    transformation is performed. Defaults to the original column names.}
}
\description{
\code{transform_code_combinations} was deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::transform_code_combinations} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
The old help pages can be found at \code{help("transform_code_combinations-deprecated")}.
    Information on deprecated functions can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
