% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_MT_omrader.R, R/copy_MT_omrader.R,
%   R/read_MT_omrader.R
\name{add_MT_omrader}
\alias{add_MT_omrader}
\alias{copy_MT_omrader}
\alias{read_MT_omrader}
\title{Manage translation from komnr to MT-avdeling and MT-region}
\usage{
add_MT_omrader(
  data,
  translation_table = komnr_2_MT_omrader,
  code_column = c("komnr"),
  new_column = c("MT_avdelingnr", "MT_avdeling", "MT_regionnr", "MT_region"),
  position = "right",
  overwrite = FALSE
)

copy_MT_omrader(
  filename = list("komnr_2_MT_avdeling.csv", "MT_omrader.csv"),
  from_path = base::paste0(set_dir_NVI("GrunndataLand"), "FormaterteData/"),
  to_path = NULL
)

read_MT_omrader(
  filename = list("komnr_2_MT_avdeling.csv", "MT_omrader.csv"),
  from_path = base::paste0(set_dir_NVI("GrunndataLand"), "FormaterteData/")
)
}
\arguments{
\item{data}{Data frame with data with a column with a komnr}

\item{translation_table}{Data frame with the table for translating from komnr to MT_areas}

\item{code_column}{The column with the coded value. Valid values are one of c("komnr", "MT_avdelingnr", "MT_regionnr"). If the column in
data has another name, it can be input as a named vector, see examples.}

\item{new_column}{The new columns that should be included into the data frame. The new columns can be up to
c("MT_avdelingnr", "MT_avdeling", "MT_regionnr", "MT_region") depending on
\code{code_column}. If the new columns in the result data frame
should have other names, \code{new_column} can be input as a named vector, see examples.}

\item{position}{[\code{character(1)}]. The position for the new columns, can 
be one of c("first", "left", "right", "last", "keep"). The input can be 
abbreviated, but must be unique, i.e. c("f", "le", "r", "la", "k").}

\item{overwrite}{[\code{logical(1)}]. When the new column(s) already exist, 
the content in the existing column(s) is replaced by new data if 
\code{overwrite = TRUE}. If the new column(s) already exist and 
\code{overwrite = FALSE}, an error is issued. Defaults to \code{FALSE}.}

\item{filename}{a list with the filenames of the source files with the tables for generating the translation table.}

\item{from_path}{Path for the source files for the translation table.}

\item{to_path}{Path to where the source files for the translation table should be copied.}
}
\value{
\code{add_MT_omrader} A data frame where the MT_avdelingnr has been added in the column to the
    right of the column with the komnr.

    \code{read_MT_omrader} A data frame with the table for translating from komnr to
    c("MT_avdelingnr", "MT_avdeling", "MT_regionnr", "MT_region") as read from the source
    csv file. If not changing standard input to the function, the standard files at
    NVI's internal network is read.

    \code{copy_MT_omrader} Copies the csv-files "komnr_2_MT_avdeling.csv" and "MT_omrader.csv"
    to another directory. If the target files already exists the source files are only
    copied if they are newer than the target files.
}
\description{
Function to add columns with MT_avdelingnr, MT_avdelng (name), MT_regionnr and MT_region (name). In addition there are
   functions to read and copy the translation tables.
}
\details{
\code{add_MT_omrader} can be used to translate the komnr into MT_avdelingnr, MT_avdeling, MT_regionnr and MT_region. The
    function can also be used to translate MT_avdelingnr into MT_avdeling, MT_regionnr and MT_region or to translate MT_regionnr into
    MT_region. When the \code{code_column =} in the dataframe is not equal to one of c("komnr", "MT_avdelingnr", "MT_regionnr") the
    \code{code_column =} can be input as a named vector. Likewise, if the new columns should be given other names than
    c("MT_avdelingnr", "MT_avdeling", "MT_regionnr", "MT_region"), the \code{new_column =} can be input as a named vector, see examples.

    \code{add_MT_omrader} uses a premade translation table (komnr_2_MT_avdeling.csv). These data need to be loaded by \code{read_MT_omrader}
    before running \code{add_MT_omrader}, see example. "komnr_2_MT_avdeling.csv" is made based on information in PJS adresseregister.
    The translation table is updated when we know there is a need.

    \code{position =} is used to give the place if the new columns in the data.frame. For \code{position = "right"} the new variables are
    placed to the right of the code_variable. Likewise, for \code{position = "left"} the new variables are placed to the left of the
    code_variable. If \code{position = "first"} or \code{position = "last"} the new columns are placed first or last, respectively, in the
    data.frame. A special case occurs for \code{position = "keep"} which only has meaning when the new column has the same name as an existing
    column and overwrite = TRUE. In these cases, the existing column will be overwritten with new data and have the same position.

    \code{read_MT_omrader} reads the files "komnr_2_MT_avdeling.csv" and "MT_omrader.csv" into a data frame, usually named komnr_2_MT_omrader.
    This file is used by \code{add_MT_omrader}. If no options to the function is given, the function will read the latest updated files
    from NVI's internal network. If changing the \code{from_path}, the function can be used to read the translation file from other directories.
    This can be useful if having a script that don't have access to NVI's internal network.

    \code{copy_MT_omrader} Copies the csv-files "komnr_2_MT_avdeling.csv" and "MT_omrader.csv" to another directory. Thereby, these files are
    available for \code{read_MT_omrader} if they should be read from another directory.
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
komnr_2_MT_omrader <- read_MT_omrader()

# Copy the csv-files used to generate the translation table from the standard location to
# the subdirectory Data below the working directory
copy_MT_omrader(to_path = "./Data/")

# Reading from the subdirectory Data below the working directory
komnr_2_MT_omrader <- read_MT_omrader(from_path = "./Data/")

# Add new columns with MT_avdelingnr, MT_avdeling, MT_regionnr, and MT_region based on komnr
# Remember to load "komnr_2_MT_omrader" by "read_MT_omrader()" before running "add_MT_omrader",
# see above.
newdata <- add_MT_omrader(olddata,
                       translation_table = list(komnr_2_MT_omrader),
                       code_column = "komnr",
                       new_column = c("MT_avdelingnr", "MT_avdeling", "MT_regionnr", "MT_region"))

# Add new columns with MT_avdelingnr and MT_avdeling based on komnr. The colname of the column
# with komnr is komnr and the new columns are renamed to MT_avdnr and MT_avd.
# Remember to load "komnr_2_MT_omrader" by "read_MT_omrader()" before running "add_MT_omrader",
# see above.
newdata <- add_MT_omrader(olddata,
                       translation_table = list(komnr_2_MT_omrader),
                       code_column = c("komnr" = "komnr"),
                       new_column = c("MT_avdnr" = "MT_avdelingnr", "MT_avd" = "MT_avdeling"))

# Add new columns with MT_region based on MT_regionnr. MT_region is renamed to MT_regionnavn
# Remember to load "komnr_2_MT_omrader" by "read_MT_omrader()" before running "add_MT_omrader",
# see above.
newdata <- add_MT_omrader(olddata,
                       translation_table = list(komnr_2_MT_omrader),
                       code_column = "MT_region",
                       new_column = c("MT_regionnavn" = "MT_region"))
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
