% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_PJS_code_description-deprecated.R
\name{add_PJS_code_description}
\alias{add_PJS_code_description}
\alias{copy_PJS_codes_2_text}
\alias{read_PJS_codes_2_text}
\title{add_PJS_code_description is Deprecated}
\usage{
add_PJS_code_description(
  data,
  translation_table = PJS_codes_2_text,
  PJS_variable_type,
  code_colname,
  new_column,
  position = "right",
  overwrite = FALSE,
  backward = FALSE,
  impute_old_when_missing = FALSE
)

copy_PJS_codes_2_text(
  filename = "PJS_codes_2_text.csv",
  from_path = paste0(set_dir_NVI("Provedata_Rapportering"), "FormaterteData/"),
  to_path = NULL
)

read_PJS_codes_2_text(
  filename = "PJS_codes_2_text.csv",
  from_path = paste0(set_dir_NVI("Provedata_Rapportering"), "FormaterteData/")
)
}
\arguments{
\item{data}{[\code{data.frame}] \cr
PJS data with at least one column that have codes for a PJS variable.}

\item{translation_table}{[\code{data.frame}] \cr
Table with the code and the description for PJS variables. Defaults to
"PJS_codes_2_text".}

\item{PJS_variable_type}{[\code{character}] \cr
One or more PJS variables, for example "hensikt". See details for a list
of all PJS variables included in the pre made translation table
"pjscode_2_descriptions.csv". If more than one code type should be translated,
they can be given in the vector. You may also use argument
\code{PJS_variable_type = "auto"}, if \code{code_colname} have standardized
PJS column names only, see details.}

\item{code_colname}{[\code{character}] \cr
The name of the column with codes that should be translated. If several codes
should be translated, a vector with the names of the coded variables should be given.}

\item{new_column}{[\code{character}] \cr
The name of the new column with the text describing the code. If several
codes should be translated, a vector with the new column names should be
given. You may also use argument \code{new_column = "auto"}, if \code{code_colname}
have standardized PJS column names only, see details.}

\item{position}{[\code{character}] \cr
Position for the new columns, can be one of c("first", "left", "right",
"last", "keep"). If several codes should be translated, either one value
to be applied for all may be given or a vector with specified position
for each code to be translated should be given. Defaults to "right".}

\item{overwrite}{[\code{logical(1)}]\cr
When the new column(s) already exist, 
    the content in the existing column(s) is replaced by new data if 
    \code{overwrite = TRUE}. If the new column(s) already exist and 
    \code{overwrite = FALSE}, an error is issued. Defaults to \code{FALSE}.}

\item{backward}{[\code{logical(1)}] \cr
If \code{TRUE}, it translates from descriptive text and back to PJS code,
see details. Defaults to \code{FALSE}.}

\item{impute_old_when_missing}{[\code{logical(1)}] \cr
Should existing value be transferred if no value for the code is found?
Defaults to \code{FALSE}.}

\item{filename}{[\code{character(1)}] \cr
File name of the source file for the translation table for PJS codes.}

\item{from_path}{[\code{character(1)}] \cr
Path for the source translation table for PJS codes.}

\item{to_path}{[\code{character(1)}] \cr
Path for the target translation table for PJS codes when copying the
translation table.}
}
\description{
\code{add_PJS_code_description}, \code{read_PJS_code_2_text} and
    \code{read_PJS_code_2_text} were deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::add_PJS_code_description}, \code{NVIpjsr::read_PJS_code_2_text}
    and \code{NVIpjsr::read_PJS_code_2_text} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
The old help pages can be found at \code{help("add_PJS_code_description-deprecated")}.
    Information on deprecated function can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
