% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_PJS_code_description.R,
%   R/copy_PJS_code_2_text.R, R/read_PJS_code_2_text.R
\name{add_PJS_code_description}
\alias{add_PJS_code_description}
\alias{copy_PJS_codes_2_text}
\alias{read_PJS_codes_2_text}
\title{Manage translation of PJS-codes to descriptive text}
\usage{
add_PJS_code_description(
  data,
  translation_table = PJS_codes_2_text,
  PJS_variable_type,
  code_colname,
  new_column,
  position = "right",
  overwrite = FALSE
)

copy_PJS_codes_2_text(
  filename = "PJS_codes_2_text.csv",
  from_path = paste0(set_dir_NVI("Provedata_Rapportering"), "FormaterteData/"),
  to_path = NULL
)

read_PJS_codes_2_text(
  filename = "PJS_codes_2_text.csv",
  from_path = paste0(set_dir_NVI("Provedata_Rapportering"), "FormaterteData/")
)
}
\arguments{
\item{data}{Data frame with PJS-data with a column with codes for a PJS-variable}

\item{translation_table}{Data frame with the code and the description for PJS-variables}

\item{PJS_variable_type}{A vector with PJS-variables, for example "hensikt". See details for a list of all PJS-variables included in
the premade translation table pjscode_2_descriptions.csv. If more than one code should be translated, they can be given in the vector.}

\item{code_colname}{The name of the column with codes that should be translated. If several codes should be translated, a vector with the
names of the coded variables should be given.}

\item{new_column}{The name of the new column with the text describing the code. If several codes should be translated, a vector with the
new column names should be given.}

\item{position}{position for the new columns, can be one of c("first", "left", "right", "last", "keep"). If several codes should be translated,
either one value to be applied for all may be given or a vector with specified position for each code to be translated should be given.}

\item{overwrite}{When the new column(s) already exist, the content in the existing column(s) is replaced by new data if overwrite = TRUE.
If the new columns already exists and overwrite = FALSE, an error will be issued.}

\item{filename}{Filename of the source file for the translation table for PJS-codes}

\item{from_path}{Path for the source translation table for PJS-codes}

\item{to_path}{Path for the target translation table for PJS-codes when copying the table}
}
\value{
\code{add_PJS_code_description} A dataframe where the description text for the PJS-code has been added in the column to the
    right of the column with the code. If the input is a tibble, it will be transformed to a dataframe.

    \code{read_PJS_codes_2_text} A dataframe with the translation table for PJS-codes as read from the source csv-file. If not changing standard
    input, the standard file at NVI's internal network is read.

    \code{copy_PJS_codes_2_text} Copies the source translation table for PJS-codes to another location. If the target file already exists
    the source file is only copied if it is newer than the target file.
}
\description{
Functions to adds a column with descriptive text for a column with PJS-codes in a data frame with PJS-data.
   In addition there are functions to read and copy an updated version of the PJS code registers.
}
\details{
Export of data from PJS will produce data frames in which many columns have coded data. These need to be translated
    into descriptive text to increase readability.

    \code{add_PJS_code_description} can be used to translate the codes into descriptive text. In a data.frame with coded values,
    the function can return a data frame with the descriptive text in a new column. As default, the descriptive text is input 
    in a new column to the right of the column with codes.

    \code{add_PJS_code_description} uses a premade translation table (PJS_codes_2_text.csv). The data need to be loaded by 
    \code{read_PJS_code_registers} before running \code{add_PJS_code_description}, see example. The file "PJS_codes_2_text.csv" is 
    normally updated every night from PJS.

    Currently, the translation table has PJS-codes and the corresponding description for the following PJS variable types:

\tabular{llll}{
  \strong{PJS variable type} \tab \strong{code colname} \tab \strong{new column} \tab \strong{remark} \cr
  seksjon \tab ansvarlig_seksjon \tab ansvarlig_seksjon_navn \tab \cr
  hensikt \tab hensiktkode \tab hensikt \tab \cr
  utbrudd \tab utbruddnr \tab utbrudd \tab translates NVI's outbreak number \cr
  registertype \tab rekvirenttype \tab rekvirenttypen \tab categories of locations and addresses \cr
  registertype \tab eier_lokalitettype \tab eier_lokalitettypen \tab categories of locations and addresses \cr
  registertype \tab annen_aktortype \tab annen_aktortypen \tab categories of locations and addresses \cr
  art \tab artkode \tab art \tab species and breed codes to species name \cr
  fysiologisk_stadium \tab fysiologisk_stadiumkode \tab fysiologisk_stadium \tab \cr
  kjonn \tab kjonn \tab kjonn_navn \tab \cr
  driftsform \tab driftsformkode \tab driftsform \tab \cr
  oppstalling \tab oppstallingkode \tab oppstalling \tab \cr
  provetype \tab provetypekode \tab provetype \tab \cr
  provemateriale \tab provematerialekode \tab provemateriale \tab \cr
  forbehandling \tab forbehandlingkode \tab forbehandling \tab \cr
  metode \tab metodekode \tab metode \tab \cr
  konkl_type \tab konkl_type \tab konkl_type_navn \tab \cr
  kjennelse \tab konkl_kjennelsekode \tab konkl_kjennelse \tab \cr
  kjennelse \tab res_kjennelsekode \tab res_kjennelse \tab \cr
  analytt \tab konkl_analyttkode \tab konkl_analytt \tab \cr
  analytt \tab res_analyttkode \tab res_analytt \tab \cr
}

In addition artrase (species and breed codes to species or breed name)
    \code{position =} is used to give the place if the new columns in the data.frame. For \code{position = "right"} the new variables are
    placed to the right of the code_variable. Likewise, for \code{position = "left"} the new variables are placed to the left of the
    code_variable. If \code{position = "first"} or \code{position = "last"} the new columns are placed first or last, respectively, in the
    data.frame. A special case occurs for \code{position = "keep"} which only has meaning when the new column has the same name as an existing
    column and overwrite = TRUE. In these cases, the existing column will be overwritten with new data and have the same position.

    \code{read_PJS_code_registers} reads the file "PJS_codes_2_text.csv" into a dataframe that can be used by \code{add_PJS_code_description}.
    In standard setting will the file read in the latest updated file from NVI's internal network. If changing the \code{from_path}, the function can
    be used to read the translation file from other directories. This can be useful if having a stand alone app with no connection the NVI's
    internal network. In other cases, it should be avoided.

    PJS_codes_2_text.csv has the following columns: c("type", "kode", "navn", "utgatt_dato"), where "type" is the PJS variable type
    as listed above (for example hensikt), "kode" is the variable with the PJS-code, "navn" is the text describing the code, and "utgatt_dato"
    is the date for last date that the code was valid (NA if still valid). If translation tables are needed for other PJS variables, a dataframe
    with the same definition can be constructed to translate new variables.

    \code{copy_PJS_code_registers} copies the file pjsCodeDescriptions.csv to a chosen directory.
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
PJS_codes_2_text <- read_PJS_codes_2_text()

# Copy standard file from standard location to the subdirectory Data below the working directory
copy_PJS_codes_2_text(to_path = "./Data/")

# Reading from the subdirectory Data below the working directory
PJS_codes_2_text <- read_PJS_codes_2_text("PJS_codes_2_text.csv", "./Data/")

# Translating artkode into art
newdata <- add_PJS_code_description(olddata, PJS_codes_2_text, "art", "artkode", "art")

# Translating hensiktkode into Hensikt and konklusjonkode to Konklusjonskjennelse
newdata <- add_PJS_code_description(olddata,
                                    PJS_codes_2_text,
                                    PJS_variable_type = c("hensikt", "kjennelse"),
                                    code_colname = c("hensiktkode", "konklusjonkode"),
                                    new_column = c("Hensikt", "Konklusjonskjennelse"))
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
