% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_query_outbreak-deprecated.R
\name{build_query_outbreak-defunct}
\alias{build_query_outbreak-defunct}
\title{Builds query to select data for a disease outbreak from PJS}
\arguments{
\item{period}{[\code{numeric}]\cr
Time period given as year. One year or a vector giving the first
and last years that should be selected.}

\item{utbrudd}{[\code{character}]\cr
Utbruddsid(er) that should be selected. Defaults to \code{NULL}.}

\item{hensikt}{[\code{character}]\cr
Specific hensiktkoder. If sub-hensikter should be included,
end the code with \%. Defaults to \code{NULL}.}

\item{analytt}{[\code{character}]\cr
Analyttkoder that should be selected. If sub-analytter should be included,
end the code with \%. Defaults to \code{NULL}.}

\item{metode}{[\code{character}]\cr
Specific metodekoder. Defaults to \code{NULL}.}

\item{db}{[\code{character(1)}]\cr
The database for which the query is built. Defaults to "PJS" that currently
is the only valid value.}
}
\value{
A list with select-statement for v2_sak_m_res and v_sakskonklusjon to
    be included in a \code{RODBC::sqlQuery}.
}
\description{
Builds a query to select all data for a disease outbreak from
    PJS. The input are utbruddsid, hensiktskoder, analyttkoder for the
    infectious agent and/or disease, and metodekoder specific for the
    infection and/or disease. The the query is written in T-SQL as used by MS-SQL.
}
\details{
The function builds the SQL syntax to select all PJS-saker
    regarding a disease outbreak from PJS. This is based on a outbreak being
    defined by an utbruddsid, hensiktkoder and/or analyttkoder for the
    infectious agent and/or disease. At least one of these must be given as
    input to the function.

    The utbruddsid is the internal id of the utbrudd in the utbrudds-register
    in PJS. One or more utbruddsid may be given as input.

    One or more hensiktkoder may be input to the selection statement. These
    may define the outbreak by themselves or may be input in addition to the
    utbruddsid and/or analyttkode.

    One or more analyttkoder may be input to the selection statement. These
    may define the outbreak by themselves or may be input in addition to the
    utbruddsid and/or hensiktkode.

    In addition one or more specific metoder may be input to the selection
    statement. With specific metode is meant a metode that implies an
    examination that will give one of the input analytter as a result. These
    cannot be sufficient to define the outbreak, but is included if the
    outbreak is defined as all samples examined for a specific analytt.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
