% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sql_modules-defunct.R
\name{build_sql_modules-defunct}
\alias{build_sql_modules-defunct}
\title{Builds sql modules to be included in select statements for PJS}
\arguments{
\item{year}{[\code{numeric}]\cr
One year or a vector giving the first and last years that should 
be selected.}

\item{values}{[\code{character}]\cr
The value of the codes that should be selected. If sub-codes should be
included, add "\%" after the code, see example.}

\item{varname}{[\code{character(1)}]\cr
The PJS variable name of the variable in PJS from which the
coded values should be selected.}

\item{db}{[\code{character(1)}]\cr
The database for which the query is built. Defaults to "PJS" that currently
is the only valid value.}
}
\value{
SQL-code to be included when building select-statements for PJS.
}
\description{
Builds sql modules to be included in select statements for PJS
   when building queries for selecting data. The functions takes the values
   for which observations should be selected as input and builds the
   sql syntax.

\code{build_sql_select_code} was deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::build_sql_select_code} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
\code{build_sql_select_year} builds the SQL syntax to select observations
    from one or more consecutive years from PJS. The input can be given as
    one year, the first and last year or a range of years. If a range is given,
    this will be interpreted as first and last years and all years in between
    will be included.

\code{build_sql_select_code} builds the SQL syntax to select observations
    with the given code values from one variable in PJS with hierarchical codes.
    When the code value including sub codes  should be selected, add "%" to the
    code, see example.

Be aware that these functions only builds an sql building block to be
    included into a select statement. It will not build a complete select
    statement. These functions are mainly intended for internal use and
    are called from
    \ifelse{html}{\code{\link{build_query_hensikt}}}{\code{build_query_hensikt}},
    \ifelse{html}{\code{\link{build_query_one_disease}}}{\code{build_query_one_disease}}
    and
    \ifelse{html}{\code{\link{build_query_outbreak}}}{\code{build_query_outbreak}}.
    If generating own select statements, these can be used to facilitate
    the coding. The building blocks can be combined with "AND" and "OR"
    and brackets to get the intended select statement.

The old help pages can be found at \code{help("build_sql_modules-deprecated")}.
    Information on deprecated functions can be found at \code{help("NVIdb-deprecated")}.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
