% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sql_modules.R
\name{build_sql_select_year}
\alias{build_sql_select_year}
\alias{build_sql_select_code}
\title{Builds sql module for selecting year from PJS}
\usage{
build_sql_select_year(year, varname, db = "PJS")

build_sql_select_code(values, varname, db = "PJS")
}
\arguments{
\item{year}{the year that should be selected as integer. Can be given as one year, the first and last year or a range of years.}

\item{varname}{The variable name of the variable from which tha coded values should be selected}

\item{db}{the database for which the query is built. Currently only the value "PJS" is accepted.}

\item{values}{the value of the codes that should be selected given as character. If sub-codes should be included, add "%"
after the code, see example.}
}
\value{
a SQL-code for selecting year from PJS to be included when building select-statements.

a SQL-code for selecting the codes from PJS to be included when building select-statements .
}
\description{
Builds a sql module for selecting one or more years from PJS. This can be included into queries for
   selecting data from PJS.

Builds a sql module for selecting one or more codes from PJS. This can be included into queries for
   selecting data from PJS.
}
\details{
The function builds the SQL syntax to select one or more consequtive years from PJS.

The function builds the SQL syntax to select one or more codes for one variable from PJS.
}
\examples{
# SQL-select module for selecting year from PJS
build_sql_select_year(year = 2020, varname = "aar")

build_sql_select_year(year = c(2019, 2021), varname = "aar")

build_sql_select_year(year = c(2019:2021), varname = "aar")

# SQL-select module for selecting hensiktkode from PJS
build_sql_select_code(values = "0100101", varname = "hensiktkode", db = "PJS") 

build_sql_select_code(values = "0100101\%", varname = "hensiktkode", db = "PJS") 

build_sql_select_code(values = c("0100101", "0100101007", "0100102\%", "0100202\%"), 
                      varname = "hensiktkode", 
                      db = "PJS") 

}
\author{
Petter Hopp Petter.Hopp@vetinst.no

Petter Hopp Petter.Hopp@vetinst.no
}
