% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIdb-defunct.R, R/login_EOS-defunct.R,
%   R/login_PJS-defunct.R, R/set_PAT-defunct.R, R/set_credentials_EOS-defunct.R
\name{NVIdb-defunct}
\alias{NVIdb-defunct}
\alias{login_EOS}
\alias{login_by_input_EOS}
\alias{login_by_credentials_EOS}
\alias{login_PJS}
\alias{login_by_input_PJS}
\alias{login_by_credentials_PJS}
\alias{set_PAT}
\alias{get_PAT}
\alias{remove_PAT}
\alias{set_credentials_EOS}
\title{Defunct Functions in Package NVIdb}
\usage{
login_EOS(...)

login_by_input_EOS(...)

login_by_credentials_EOS(...)

login_PJS(...)

login_by_input_PJS(...)

login_by_credentials_PJS(...)

set_PAT(...)

get_PAT(...)

remove_PAT(...)

set_credentials_EOS(...)
}
\arguments{
\item{\dots}{(arguments)}
}
\value{
(results)
}
\description{
The functions listed in "Usage" are no longer part of NVIdb.
    When possible, alternative functions are mentioned. The original help
    pages for defunct functions are available at \code{help("<function>-defunct")}.
}
\details{
\code{set_PAT}, \code{get_PAT}, and \code{remove_PAT} were defunct from
    v0.14.1 released 2025-10-## and had been deprecated from v0.11.0 released
    2024-01-24. The functions were never taken into use. Functions from the
    much better package \code{gitcreds} should be used instead.

The wrapper function \code{set_credentials_EOS} was defunct from v0.14.1
    released 2025-10-## and had been deprecated from v0.11.2 released
    2024-05-03. The function was rarely used and was an unnecessary wrapper
    around \code{set_credentials}. \code{\link{set_credentials}} with the
    argument \code{dbservice} = "EOS" should be used instead. See
    \code{NVIdb-defunct} (help("NVIdb-defunct")).

The wrapper functions \code{login_EOS}, \code{login_by_input_EOS}, and
    \code{login_by_credentials_EOS} were defunct from v0.14.1 released
    2025-10-## and had been deprecated from v0.11.2 released
    2024-05-03. These functions were rarely used and were unnecessary
    wrappers around the login-functions. Use the corresponding
    login-functions with the argument \code{dbservice = "EOS"} instead.
    See \code{NVIdb-defunct} (help("NVIdb-defunct")).

The wrapper functions \code{login_PJS}, \code{login_by_input_PJS}, and
    \code{login_by_credentials_PJS} were defunct from
    v0.14.1 released 2025-10-## and had been deprecated from v0.13.0 released
    2024-12-13. These functions were unnecessary wrappers around the
    login-functions. Use the corresponding login-functions with the argument
    \code{dbservice = "PJS"} instead. See \code{NVIdb-defunct}
    (help("NVIdb-defunct")).
}
\examples{
\dontrun{
set_PAT(...) ### -- use gitcreds::gitcreds_set() instead.
get_PAT(...) ### -- use gitcreds::gitcreds_get() instead.
remove_PAT(...) ### -- use gitcreds::gitcreds_delete() instead.
set_credentials_EOS(...) ### -- use set_credentials("EOS") instead.
login_EOS(...) ### -- use login("EOS") instead.
login_by_input_EOS(...) ### -- use login_by_input("EOS") instead.
login_by_credentials_EOS(...) ### -- use login_by_credentials("EOS") instead.
login_PJS(...) ### -- use login("PJS") instead.
login_by_input_PJS(...) ### -- use login_by_input("PJS") instead.
login_by_credentials_PJS(...) ### -- use login_by_credentials("PJS") instead
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
