% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIdb-defunct.R,
%   R/add_PJS_code_description-deprecated.R, R/build_query_hensikt-defunct.R,
%   R/build_query_one_disease-defunct.R, R/build_query_outbreak-defunct.R,
%   R/build_sql_modules-defunct.R, R/choose_PJS_levels-defunct.R,
%   R/exclude_from_PJSdata-defunct.R, R/login_EOS-defunct.R,
%   R/login_PJS-defunct.R, R/read_eos_data-defunct.R,
%   R/retrieve_PJSdata-defunct.R, R/select_PJSdata_for_value-defunct.R,
%   R/set_PAT-defunct.R, R/set_credentials_EOS-defunct.R,
%   R/set_disease_parameters-defunct.R, R/standardize_PJSdata-defunct.R,
%   R/standardize_eos_data-defunct.R, R/transform_code_combinations-defunct.R
\name{NVIdb-defunct}
\alias{NVIdb-defunct}
\alias{add_PJS_code_description}
\alias{copy_PJS_codes_2_text}
\alias{read_PJS_codes_2_text}
\alias{build_query_hensikt}
\alias{build_query_one_disease}
\alias{build_query_outbreak}
\alias{build_sql_select_year}
\alias{build_sql_select_code}
\alias{choose_PJS_levels}
\alias{exclude_from_PJSdata}
\alias{login_EOS}
\alias{login_by_input_EOS}
\alias{login_by_credentials_EOS}
\alias{login_PJS}
\alias{login_by_input_PJS}
\alias{login_by_credentials_PJS}
\alias{read_eos_data}
\alias{retrieve_PJSdata}
\alias{select_PJSdata_for_value}
\alias{set_PAT}
\alias{get_PAT}
\alias{remove_PAT}
\alias{set_credentials_EOS}
\alias{set_disease_parameters}
\alias{standardize_PJSdata}
\alias{standardize_eos_data}
\alias{transform_code_combinations}
\title{Defunct Functions in Package NVIdb}
\usage{
add_PJS_code_description(...)

copy_PJS_codes_2_text(...)

read_PJS_codes_2_text(...)

build_query_hensikt(...)

build_query_one_disease(...)

build_query_outbreak(...)

build_sql_select_year(...)

build_sql_select_code(...)

choose_PJS_levels(...)

exclude_from_PJSdata(...)

login_EOS(...)

login_by_input_EOS(...)

login_by_credentials_EOS(...)

login_PJS(...)

login_by_input_PJS(...)

login_by_credentials_PJS(...)

read_eos_data(...)

retrieve_PJSdata(...)

select_PJSdata_for_value(...)

set_PAT(...)

get_PAT(...)

remove_PAT(...)

set_credentials_EOS(...)

set_disease_parameters(...)

standardize_PJSdata(...)

standardize_eos_data(...)

transform_code_combinations(...)
}
\arguments{
\item{\dots}{(arguments)}
}
\value{
(results)
}
\description{
The functions listed in "Usage" are no longer part of NVIdb.
    When possible, alternative functions are mentioned. The original help
    pages for defunct functions are available at \code{help("<function>-defunct")}.
}
\details{
\code{set_PAT}, \code{get_PAT}, and \code{remove_PAT} were defunct from
    v0.14.1 released 2025-10-29 and had been deprecated from v0.11.0 released
    2024-01-24. The functions were never taken into use. Functions from the
    much better package \code{gitcreds} should be used instead.

The wrapper function \code{set_credentials_EOS} was defunct from v0.14.1
    released 2025-10-29 and had been deprecated from v0.11.2 released
    2024-05-03. The function was rarely used and was an unnecessary wrapper
    around \code{set_credentials}. \code{\link{set_credentials}} with the
    argument \code{dbservice} = "EOS" should be used instead.

The wrapper functions \code{login_EOS}, \code{login_by_input_EOS}, and
    \code{login_by_credentials_EOS} were defunct from v0.14.1 released
    2025-10-29 and had been deprecated from v0.11.2 released
    2024-05-03. These functions were rarely used and were unnecessary
    wrappers around the login-functions. Use the corresponding
    login-functions with the argument \code{dbservice = "EOS"} instead.

The wrapper functions \code{login_PJS}, \code{login_by_input_PJS}, and
    \code{login_by_credentials_PJS} were defunct from
    v0.14.1 released 2025-10-29 and had been deprecated from v0.13.0 released
    2024-12-13. These functions were unnecessary wrappers around the
    login-functions. Use the corresponding login-functions with the argument
    \code{dbservice = "PJS"} instead.

The functions \code{read_eos_data} and \code{standardize_eos_data} have been
    moved to \code{NVIpjsr}. These were defunct from v0.14.1 released
    2025-10-29 and had been deprecated from v0.13.0 released 2024-12-13. Use
    the corresponding functions in \code{NVIpjsr}.

The functions \code{add_PJS_code_description}, \code{copy_PJS_code_2_text},
    \code{read_PJS_code_2_text}, \code{build_query_hensikt},
    \code{build_query_one_disease}, \code{build_query_one_outbreak},
    \code{build_sql_select_code}, \code{build_sql_select_year},
    \code{choose_PJS_levels}, \code{exclude_from_PJSdata},
    \code{retrieve_PJSdata}, \code{select_PJSdata_for_value},
    \code{set_disease_parameters}, \code{standardize_PJSdata}, and
    \code{transform_code_combinations}
    have been moved to \code{NVIpjsr}. These were defunct from v0.15.1 released
    2026-##-## and had been deprecated from v0.13.0 released 2024-12-13. Use
    the corresponding functions in \code{NVIpjsr}.
}
\examples{
\dontrun{
set_PAT(...) ### -- use gitcreds::gitcreds_set() instead.
get_PAT(...) ### -- use gitcreds::gitcreds_get() instead.
remove_PAT(...) ### -- use gitcreds::gitcreds_delete() instead.
set_credentials_EOS(...) ### -- use set_credentials("EOS") instead.
login_EOS(...) ### -- use login("EOS") instead.
login_by_input_EOS(...) ### -- use login_by_input("EOS") instead.
login_by_credentials_EOS(...) ### -- use login_by_credentials("EOS") instead.
login_PJS(...) ### -- use login("PJS") instead.
login_by_input_PJS(...) ### -- use login_by_input("PJS") instead.
login_by_credentials_PJS(...) ### -- use login_by_credentials("PJS") instead
read_eos_data(...) ### -- use NVIpjsr::read_eos_data() instead
standardize_eos_data(...) ### -- use NVIpjsr::standardize_eos_data() instead
add_PJS_code_description(...) ### -- use NVIpjsr::add_PJS_code_description() instead
copy_PJS_code_2_text(...) ### -- use NVIpjsr::copy_PJS_code_2_text() instead
read_PJS_code_2_text(...) ### -- use NVIpjsr::read_PJS_code_2_text() instead
build_query_hensikt(...) ### -- use NVIpjsr::build_query_hensikt() instead
build_query_one_disease(...) ### -- use NVIpjsr::build_query_one_disease() instead
build_query_outbreak(...) ### -- use NVIpjsr::build_query_outbreak() instead
build_sql_select_code(...) ### -- use NVIpjsr::build_sql_select_code() instead
build_sql_select_year(...) ### -- use NVIpjsr::build_sql_select_year() instead
choose_PJS_levels(...) ### -- use NVIpjsr::choose_PJS_levels() instead
exclude_from_PJSdata(...) ### -- use NVIpjsr::exclude_from_PJSdata() instead
retrieve_PJSdata(...) ### -- use NVIpjsr::retrieve_PJSdata() instead
select_PJSdata_for_value(...) ### -- use NVIpjsr::select_PJSdata_for_value() instead
set_disease_parameters(...) ### -- use NVIpjsr::set_disease_parameters() instead
standardize_PJSdata(...) ### -- use NVIpjsr::standardize_PJSdata() instead
transform_code_combinations(...) ### -- use NVIpjsr::transform_code_combinations() instead
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{!=}
\keyword{!is.na(PJS_types_selected$type))}
\keyword{!substr(code_2_description$kode,}
\keyword{"'}
\keyword{",}
\keyword{"-"}
\keyword{"05")}
\keyword{"NVIdb",}
\keyword{"NVIpjsr",}
\keyword{"You}
\keyword{"`"}
\keyword{"`.}
\keyword{"add_PJS_code_description",}
\keyword{"antall")}
\keyword{"art"}
\keyword{"art")}
\keyword{"auto"}
\keyword{"auto")}
\keyword{"auto".}
\keyword{"code_colname")}
\keyword{"installed")))}
\keyword{"keep"),}
\keyword{"kode")}
\keyword{"kode")),}
\keyword{"kode"]}
\keyword{"kode"],}
\keyword{"last",}
\keyword{"left",}
\keyword{"named",}
\keyword{"navn")}
\keyword{"right",}
\keyword{"the}
\keyword{#}
\keyword{%in%}
\keyword{&}
\keyword{'",}
\keyword{'NVIpjsr::add_PJS_code_description'"))}
\keyword{'PJS_variable_type'}
\keyword{'auto'",}
\keyword{'new_column'}
\keyword{(PJS_variable_type[1]}
\keyword{(PJS_variable_type[i]}
\keyword{(i}
\keyword{(isFALSE(overwrite))}
\keyword{(isTRUE(NVIcheckmate::check_package(x}
\keyword{(isTRUE(backward)}
\keyword{(isTRUE(backward))}
\keyword{(isTRUE(impute_old_when_missing_backward))}
\keyword{(length(position)}
\keyword{(new_column[1]}
\keyword{(txtvarname)}
\keyword{),}
\keyword{,}
\keyword{-}
\keyword{----}
\keyword{.Deprecated(new}
\keyword{.data$kode)}
\keyword{.data$kode,}
\keyword{.data$navn,}
\keyword{1}
\keyword{1)}
\keyword{1,}
\keyword{11)}
\keyword{1:length(code_colname))}
\keyword{2)}
\keyword{8)}
\keyword{<-}
\keyword{=}
\keyword{==}
\keyword{>}
\keyword{ARGUMENT}
\keyword{ARGUMENTS}
\keyword{CHECKING}
\keyword{Calls}
\keyword{Changes}
\keyword{DEPRECATED}
\keyword{Either}
\keyword{FALSE}
\keyword{FALSE,}
\keyword{FUN}
\keyword{Fix}
\keyword{Generates}
\keyword{ID_column}
\keyword{ID_column_translation_table}
\keyword{Imputes}
\keyword{In}
\keyword{Make}
\keyword{NA}
\keyword{NULL}
\keyword{NVIcheckmate::assert_names(code_colname,}
\keyword{NVIcheckmate::assert_names(new_column,}
\keyword{NVIcheckmate::assert_subset_character(code_colname,}
\keyword{NVIcheckmate::assert_subset_character(x}
\keyword{NVIdb::PJS_code_description_colname}
\keyword{NVIdb::standardize_PJSdata}
\keyword{NVIpjsr::add_PJS_code_description(data}
\keyword{Object}
\keyword{PJS}
\keyword{PJS_types_selected}
\keyword{PJS_types_selected$new_column}
\keyword{PJS_types_selected$type}
\keyword{PJS_variable_type}
\keyword{PJS_variable_type,}
\keyword{PJS_variable_type/new_column}
\keyword{PJS_variable_type[i]}
\keyword{PJS_variable_type[i]),}
\keyword{PREPARE}
\keyword{Perform}
\keyword{Removes}
\keyword{Report}
\keyword{Restores}
\keyword{Should}
\keyword{Swips}
\keyword{TRUE}
\keyword{Transform}
\keyword{Transforms}
\keyword{]}
\keyword{a}
\keyword{above}
\keyword{actual}
\keyword{add}
\keyword{add_new_column(data,}
\keyword{add_new_column,}
\keyword{adds}
\keyword{after}
\keyword{afterwards.}
\keyword{already}
\keyword{and}
\keyword{any}
\keyword{any.missing}
\keyword{argument}
\keyword{as}
\keyword{as.data.frame(code_colname)}
\keyword{at}
\keyword{auto}
\keyword{backward}
\keyword{backward,}
\keyword{bakcward}
\keyword{base::colnames(code_2_description)[base::which(base::colnames(code_2_description)=="navn")]}
\keyword{base::setdiff(code_colname,}
\keyword{be}
\keyword{because}
\keyword{been}
\keyword{breeds}
\keyword{but}
\keyword{by}
\keyword{c("03",}
\keyword{c("03100203")),}
\keyword{c("first",}
\keyword{c("kode"),}
\keyword{c("kode",}
\keyword{c("navn"),}
\keyword{called}
\keyword{can}
\keyword{cannot}
\keyword{case}
\keyword{case.}
\keyword{check-results}
\keyword{checked}
\keyword{checking}
\keyword{checkmate::assert_data_frame(data,}
\keyword{checkmate::assert_data_frame(translation_table,}
\keyword{checkmate::assert_flag(backward,}
\keyword{checkmate::assert_flag(impute_old_when_missing,}
\keyword{checkmate::assert_flag(overwrite,}
\keyword{checkmate::assert_subset(PJS_variable_type,}
\keyword{checkmate::assert_vector(code_colname,}
\keyword{checkmate::assert_vector(new_column,}
\keyword{checkmate::makeAssertCollection()}
\keyword{checkmate::reportAssertions(checks)}
\keyword{checks}
\keyword{checks)}
\keyword{choices}
\keyword{code}
\keyword{code_2_description}
\keyword{code_2_description$antall}
\keyword{code_2_description$navn}
\keyword{code_2_description[which(substr(code_2_description$kode,}
\keyword{code_colname}
\keyword{code_colname),}
\keyword{code_colname,}
\keyword{code_colname[i],}
\keyword{code_colname[i]]}
\keyword{code_colname[i]])),}
\keyword{code_colname[i]],}
\keyword{code_colnames}
\keyword{code_description_colname}
\keyword{codes}
\keyword{colnames(code_2_description)}
\keyword{colnames(data)),}
\keyword{colnames(data),}
\keyword{column}
\keyword{column(s)}
\keyword{column_names}
\keyword{column_names[c(navn_nr,}
\keyword{comment}
\keyword{content"),}
\keyword{correct.}
\keyword{data}
\keyword{data"}
\keyword{data"),}
\keyword{data",}
\keyword{data$code_colname_org_case}
\keyword{data,}
\keyword{data[,}
\keyword{data[which(is.na(data[,}
\keyword{deparse(substitute(data)),}
\keyword{description}
\keyword{df}
\keyword{difficulties}
\keyword{disjunct.from}
\keyword{done}
\keyword{dplyr::add_count(.data$type,}
\keyword{dplyr::add_count(dplyr::across(c("type",}
\keyword{dplyr::distinct()}
\keyword{dplyr::filter(.data$antall}
\keyword{dplyr::filter(is.na(.data$utgatt_dato))}
\keyword{dplyr::left_join(code_description_colname,}
\keyword{dplyr::mutate(navn}
\keyword{dplyr::rename(code_description_colname,}
\keyword{dplyr::rename(kode}
\keyword{dplyr::select(-.data$antall)}
\keyword{else}
\keyword{elsewhere}
\keyword{equals}
\keyword{exist}
\keyword{existing}
\keyword{file}
\keyword{fixed}
\keyword{for}
\keyword{from}
\keyword{function}
\keyword{generated}
\keyword{give}
\keyword{have}
\keyword{i.e.}
\keyword{if}
\keyword{imputation}
\keyword{impute_old_when_missing}
\keyword{impute_old_when_missing)}
\keyword{impute_old_when_missing_backward}
\keyword{in}
\keyword{internal}
\keyword{intersect(colnames(data),}
\keyword{is}
\keyword{is.na(translation_table$utgatt_dato)),}
\keyword{isTRUE(impute_old_when_missing))}
\keyword{it}
\keyword{kode}
\keyword{kode_nr}
\keyword{kode_nr)]}
\keyword{length}
\keyword{length(code_colname)}
\keyword{length(code_colname))}}
\keyword{levels}
\keyword{list(code_2_description),}
\keyword{lower}
\keyword{min.len}
\keyword{missing}
\keyword{msg}
\keyword{must}
\keyword{name}
\keyword{name(s)}
\keyword{name(s):}
\keyword{name,}
\keyword{names}
\keyword{names.",}
\keyword{navn}
\keyword{navn_nr}
\keyword{new}
\keyword{new_colname}
\keyword{new_column}
\keyword{new_column)}
\keyword{new_column),}
\keyword{new_column,}
\keyword{new_column[1]}
\keyword{new_column[i],}
\keyword{new_column[i]]}
\keyword{new_column[i]])),}
\keyword{non-unique}
\keyword{not}
\keyword{number}
\keyword{of}
\keyword{old}
\keyword{only}
\keyword{or}
\keyword{original}
\keyword{overwrite}
\keyword{overwrite,}
\keyword{package}
\keyword{paste("'add_PJS_code_description'}
\keyword{paste("when}
\keyword{paste0("The}
\keyword{paste0("You}
\keyword{performed}
\keyword{position}
\keyword{position,}
\keyword{position[i],}
\keyword{probably}
\keyword{relation}
\keyword{relevant}
\keyword{rep(position,}
\keyword{replace}
\keyword{replaced}
\keyword{restored,}
\keyword{return(data)}
\keyword{runs}
\keyword{same}
\keyword{sapply(data[,}
\keyword{setdiff(colnames(data),}
\keyword{several}
\keyword{should}
\keyword{specify}
\keyword{standardize."),}
\keyword{standardized}
\keyword{store}
\keyword{subset}
\keyword{subset(PJS_types_selected,}
\keyword{subset(code_2_description,}
\keyword{subset.of}
\keyword{substr(code_2_description[which(substr(code_2_description$kode,}
\keyword{table}
\keyword{text}
\keyword{text,}
\keyword{that}
\keyword{the}
\keyword{time}
\keyword{to}
\keyword{to_column_translation_table}
\keyword{tolower(.data$navn))}
\keyword{tolower(code_2_description$navn)}
\keyword{tolower)}
\keyword{translated}
\keyword{translating}
\keyword{translation}
\keyword{translation,}
\keyword{translation_table}
\keyword{translation_table,}
\keyword{translation_table[base::which(translation_table$type}
\keyword{translation_tables}
\keyword{type}
\keyword{unique(code_2_description)}
\keyword{unique(code_description_colname$code_colname),}
\keyword{unique(position),}
\keyword{unique(translation_table$type),}
\keyword{use}
\keyword{used}
\keyword{usually}
\keyword{values}
\keyword{variabel}
\keyword{variables}
\keyword{vector}
\keyword{wanted}
\keyword{was}
\keyword{which(column_names}
\keyword{with}
\keyword{without}
\keyword{{}
\keyword{{position}
\keyword{|}
\keyword{|>}
\keyword{}}
