% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIdb-defunct.R,
%   R/build_query_hensikt-defunct.R, R/build_query_one_disease-defunct.R,
%   R/build_query_outbreak-defunct.R, R/choose_PJS_levels-defunct.R,
%   R/exclude_from_PJSdata-defunct.R, R/login_EOS-defunct.R,
%   R/login_PJS-defunct.R, R/read_eos_data-defunct.R,
%   R/retrieve_PJSdata-defunct.R, R/select_PJSdata_for_value-defunct.R,
%   R/set_PAT-defunct.R, R/set_credentials_EOS-defunct.R,
%   R/set_disease_parameters-defunct.R, R/standardize_PJSdata-defunct.R,
%   R/standardize_eos_data-defunct.R, R/transform_code_combinations-defunct.R
\name{NVIdb-defunct}
\alias{NVIdb-defunct}
\alias{build_query_hensikt}
\alias{build_query_one_disease}
\alias{build_query_outbreak}
\alias{choose_PJS_levels}
\alias{exclude_from_PJSdata}
\alias{login_EOS}
\alias{login_by_input_EOS}
\alias{login_by_credentials_EOS}
\alias{login_PJS}
\alias{login_by_input_PJS}
\alias{login_by_credentials_PJS}
\alias{read_eos_data}
\alias{retrieve_PJSdata}
\alias{select_PJSdata_for_value}
\alias{set_PAT}
\alias{get_PAT}
\alias{remove_PAT}
\alias{set_credentials_EOS}
\alias{set_disease_parameters}
\alias{standardize_PJSdata}
\alias{standardize_eos_data}
\alias{transform_code_combinations}
\title{Defunct Functions in Package NVIdb}
\usage{
build_query_hensikt(...)

build_query_one_disease(...)

build_query_outbreak(...)

choose_PJS_levels(...)

exclude_from_PJSdata(...)

login_EOS(...)

login_by_input_EOS(...)

login_by_credentials_EOS(...)

login_PJS(...)

login_by_input_PJS(...)

login_by_credentials_PJS(...)

read_eos_data(...)

retrieve_PJSdata(...)

select_PJSdata_for_value(...)

set_PAT(...)

get_PAT(...)

remove_PAT(...)

set_credentials_EOS(...)

set_disease_parameters(...)

standardize_PJSdata(...)

standardize_eos_data(...)

transform_code_combinations(...)
}
\arguments{
\item{\dots}{(arguments)}
}
\value{
(results)
}
\description{
The functions listed in "Usage" are no longer part of NVIdb.
    When possible, alternative functions are mentioned. The original help
    pages for defunct functions are available at \code{help("<function>-defunct")}.
}
\details{
\code{set_PAT}, \code{get_PAT}, and \code{remove_PAT} were defunct from
    v0.14.1 released 2025-10-29 and had been deprecated from v0.11.0 released
    2024-01-24. The functions were never taken into use. Functions from the
    much better package \code{gitcreds} should be used instead.

The wrapper function \code{set_credentials_EOS} was defunct from v0.14.1
    released 2025-10-29 and had been deprecated from v0.11.2 released
    2024-05-03. The function was rarely used and was an unnecessary wrapper
    around \code{set_credentials}. \code{\link{set_credentials}} with the
    argument \code{dbservice} = "EOS" should be used instead.

The wrapper functions \code{login_EOS}, \code{login_by_input_EOS}, and
    \code{login_by_credentials_EOS} were defunct from v0.14.1 released
    2025-10-29 and had been deprecated from v0.11.2 released
    2024-05-03. These functions were rarely used and were unnecessary
    wrappers around the login-functions. Use the corresponding
    login-functions with the argument \code{dbservice = "EOS"} instead.

The wrapper functions \code{login_PJS}, \code{login_by_input_PJS}, and
    \code{login_by_credentials_PJS} were defunct from
    v0.14.1 released 2025-10-29 and had been deprecated from v0.13.0 released
    2024-12-13. These functions were unnecessary wrappers around the
    login-functions. Use the corresponding login-functions with the argument
    \code{dbservice = "PJS"} instead.

The functions \code{read_eos_data} and \code{standardize_eos_data} have been
    moved to \code{NVIpjsr}. These were defunct from v0.14.1 released
    2025-10-29 and had been deprecated from v0.13.0 released 2024-12-13. Use
    the corresponding functions in \code{NVIpjsr}.

The functions \code{build_query_hensikt}, \code{build_query_one_disease},
    \code{build_query_one_disease},
    \code{exclude_from_PJSdata}, \code{transform_code_combinations},
    \code{choose_PJS_levels}, \code{retrieve_PJSdata}, \code{select_PJSdata_for_value},
    \code{set_disease_parameters}, \code{standardize_PJSdata},
    have been moved to \code{NVIpjsr}. These were defunct from v0.15.1 released
    2025-##-## and had been deprecated from v0.13.0 released 2024-12-13. Use
    the corresponding functions in \code{NVIpjsr}.
}
\examples{
\dontrun{
set_PAT(...) ### -- use gitcreds::gitcreds_set() instead.
get_PAT(...) ### -- use gitcreds::gitcreds_get() instead.
remove_PAT(...) ### -- use gitcreds::gitcreds_delete() instead.
set_credentials_EOS(...) ### -- use set_credentials("EOS") instead.
login_EOS(...) ### -- use login("EOS") instead.
login_by_input_EOS(...) ### -- use login_by_input("EOS") instead.
login_by_credentials_EOS(...) ### -- use login_by_credentials("EOS") instead.
login_PJS(...) ### -- use login("PJS") instead.
login_by_input_PJS(...) ### -- use login_by_input("PJS") instead.
login_by_credentials_PJS(...) ### -- use login_by_credentials("PJS") instead
read_eos_data(...) ### -- use NVIpjsr::read_eos_data() instead
standardize_eos_data(...) ### -- use NVIpjsr::standardize_eos_data() instead
build_query_hensikt(...) ### -- use NVIpjsr::build_query_hensikt() instead
build_query_one_disease(...) ### -- use NVIpjsr::build_query_one_disease() instead
build_query_outbreak(...) ### -- use NVIpjsr::build_query_outbreak() instead
exclude_from_PJSdata(...) ### -- use NVIpjsr::exclude_from_PJSdata() instead
choose_PJS_levels(...) ### -- use NVIpjsr::choose_PJS_levels() instead
retrieve_PJSdata(...) ### -- use NVIpjsr::retrieve_PJSdata() instead
select_PJSdata_for_value(...) ### -- use NVIpjsr::select_PJSdata_for_value() instead
set_disease_parameters(...) ### -- use NVIpjsr::set_disease_parameters() instead
standardize_PJSdata(...) ### -- use NVIpjsr::standardize_PJSdata() instead
transform_code_combinations(...) ### -- use NVIpjsr::transform_code_combinations() instead
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
