% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIdb-deprecated.R
\name{NVIdb-deprecated}
\alias{NVIdb-deprecated}
\title{Deprecated Functions in Package NVIdb}
\arguments{
\item{\dots}{(arguments)}
}
\value{
(results)
}
\description{
These functions are provided for compatibility with older
    versions of NVIdb only, and may be defunct as soon as the next release.
    When possible, alternative functions are mentioned. Help pages for
    deprecated functions are available at \code{help("<function>-deprecated")}.
}
\details{
\code{add_produsent} was deprecated from v0.8.0 released 2022-08-25
    as other properties
    than 'gjeldende_prodnr8' could not be included without breaking backward
    compatibility. Use \code{\link{add_produsent_properties}} instead and ensure
    to set the parameter \code{impute_old_when_missing = TRUE} when translating
    from "prodnr8" to "gjeldende_prodnr8" and set the parameter
    \code{impute_old_when_missing = FALSE} when translating from "prodnr8" to
    other properties.

The arguments \code{missing_art} and \code{file} in \code{set_disease_parameters}
    was deprecated from v0.11.0 released 2024-01-24. These arguments are
    replaced by the more meaningful \code{include_missing_art} and
    \code{selection_parameters}, respectively. If using the old arguments,
    the input will be transferred to the new arguments.

The arguments \code{FUN} and \code{selection_statement} in \code{retrieve_PJSdata}
    was deprecated from v0.11.2 released 2024-04-05. These arguments should
    instead be included in the list input to \code{selection_parameters}. If
    using the old arguments, the input will be transferred to
    \code{selection_parameters}.

All functions made especially to handle PJS data have been moved to \code{NVIpjsr}
    and were deprecated from v0.13.0 released 2024-12-13. Further development
    of these functions will only take place in \code{NVIpjsr}. The following
    functions have been moved:
\itemize{
  \item \code{add_PJS_code_description}
  \item \code{build_query_hensikt}
  \item \code{build_query_one_disease}
  \item \code{build_query_outbreak}
  \item \code{build_sql_select_code}
  \item \code{build_sql_select_year}
  \item \code{choose_PJS_levels}
  \item \code{copy_PJS_code_2_text}
  \item \code{exclude_from_PJSdata}
  \item \code{login_by_credentials_PJS}
  \item \code{read_eos_data}
  \item \code{read_PJS_code_2_text}
  \item \code{retrieve_PJSdata}
  \item \code{select_PJSdata_for_value}
  \item \code{set_disease_parameters}
  \item \code{standardize_eos_data}
  \item \code{standardize_PJSdata}
  \item \code{transform_code_combinations}
  }
Of these, \code{read_eos_data} and \code{standardize_eos_data} were defunct
  from 2025-10-29, and \code{exclude_from_PJSdata},
  \code{transform_code_combinations}, \code{choose_PJS_levels}, \code{retrieve_PJSdata},
  \code{select_PJSdata_for_value}, \code{set_disease_parameters},
  were defunct from 2026-##-##.
}
\examples{
\dontrun{
add_produsent(...) ### -- use add_produsent_properties() instead.
add_PJS_code_description(...) ### -- use NVIpjsr::add_PJS_code_description() instead
build_query_hensikt(...) ### -- use NVIpjsr::build_query_hensikt() instead
build_query_one_disease(...) ### -- use NVIpjsr:build_query_one_disease:() instead
build_query_outbreak(...) ### -- use NVIpjsr::build_query_outbreak() instead
build_sql_select_code(...) ### -- use NVIpjsr::build_sql_select_code() instead
build_sql_select_year(...) ### -- use NVIpjsr::build_sql_select_year() instead
copy_PJS_code_2_text(...) ### -- use NVIpjsr::copy_PJS_code_2_text() instead
read_PJS_code_2_text(...) ### -- use NVIpjsr::read_PJS_code_2_text() instead
standardize_PJSdata(...) ### -- use NVIpjsr::standardize_PJSdata() instead
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
