% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_csv_file}
\alias{read_csv_file}
\title{Read csv-file}
\usage{
read_csv_file(filename, from_path, options = NULL, ...)
}
\arguments{
\item{filename}{Filename of the file that should be read}

\item{from_path}{Path for the source file}

\item{options}{A list with the arguments colClasses and fileEncoding. If
\code{options = NULL} it defaults to
\code{list(colClasses = NA, fileEncoding = "UTF-8")}.}

\item{\dots}{Other arguments to be passed to \code{data.table::fread}.}
}
\value{
A data.frame with the data from the source file.
}
\description{
Reads files with data .
}
\details{
Used to read csv files with data for use in scripts.
}
\examples{
\dontrun{
read_csv_file(filename, from_path, columnclasses, fileencoding)
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
