% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_Pkode_2_text.R, R/read_Pkode_2_text.R
\name{copy_Pkode_2_text}
\alias{copy_Pkode_2_text}
\alias{read_Pkode_2_text}
\title{Manage translation table for produksjonstilskuddskoder (Pkoder)}
\usage{
copy_Pkode_2_text(
  filename = "Produksjonstilskuddskoder_UTF8.csv",
  from_path = paste0(set_dir_NVI("Prodtilskudd"), "StotteData/"),
  to_path = NULL
)

read_Pkode_2_text(
  filename = "Produksjonstilskuddskoder_UTF8.csv",
  from_path = paste0(set_dir_NVI("Prodtilskudd"), "StotteData/")
)
}
\arguments{
\item{filename}{name of the translation table, standard name is "Produksjonstilskuddskoder_UTF8.csv"}

\item{from_path}{Path for the translation table for produksjonstilskuddskoder}

\item{to_path}{Path for the target translation table when copying produksjonstilskuddskoder}
}
\value{
Data frame with the translation table for produksjonstilskuddskoder.

\code{read_Pkode_2_text} A data frame with the translation table for Pkoder to description as read from the csv file.
    If not changing standard input to the function, the standard file at NVI's internal network is read.

    \code{copy_Pkode_2_text} Copies the source translation table "Produksjonstilskuddskoder_UTF8.csv" to another location. If the target file
    already exists, the source file is only copied when it is newer than the target file.
}
\description{
Read and copy the translation table for produksjonstilskuddskoder (Pkoder).
}
\details{
The translation table for Pkoder contains the Pkode, descriptive text, unit of interest (Dyr), whether the code counts unique animals
    or not (used when summarising number of animals), and sortering to order the Pkoder. The register covers 2017 and later.

    \code{read_Pkode_2_text} reads the file "Produksjonstilskuddskoder_UTF8.csv" into a data frame. The standard settings will read
    the file from NVI's internal network. If changing the from_path, the function can be used to read the translation file from other
    directories. This can be useful if having a stand alone app with no connection the NVI's internal network. In other cases, it should
    be avoided.

    \code{copy_Pkode_2_text} copies the file Produksjonstilskuddskoder_UTF8.csv to a given location.
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
Pkode_2_text <- read_Pkode_2_text()

# Copy standard file from standard location to the sub directory Data below the working directory
copy_Pkode_2_text(to_path = "./Data/")

# Reading from the sub directory Data below the working directory
Pkode_2_text <- read_Pkode_2_text(from_path = "./Data")
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
