% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_Prodtilskudd.R, R/read_Prodtilskudd.R
\name{copy_Prodtilskudd}
\alias{copy_Prodtilskudd}
\alias{read_Prodtilskudd}
\title{Read Register for søknad om produksjonstilskudd}
\usage{
copy_Prodtilskudd(
  from_path = paste0(set_dir_NVI("Prodtilskudd"), "FormaterteData/"),
  to_path = NULL,
  Pkode_year = "last",
  Pkode_month = "both"
)

read_Prodtilskudd(
  from_path = paste0(set_dir_NVI("Prodtilskudd"), "FormaterteData/"),
  Pkode_year = "last",
  Pkode_month = "both"
)
}
\arguments{
\item{from_path}{Path for the produksjonstilskuddsregister.}

\item{to_path}{Target path for the files with the produksjonstilskuddsregister.}

\item{Pkode_year}{The year(s) from which the register should be read. Options is "last", or a vector with one or more years.}

\item{Pkode_month}{the month for which the register should be read. The options are c("05", "10", "both", "last") for Pkode_year = 2017
and c("03", "10", "both", "last") for Pkode_year >= 2018.}
}
\value{
\code{read_Prodtilskudd} One or more data frame(s) with the produksjonstilskuddsregister for each of the year and seasons selected.
    If the options Pkode_year = "last" and Pkode_month = "last" is given, one file with the last produksjonstilskuddsregister is given.

    \code{copy_Prodtilskudd} copies the source produksjonstilskuddsregister for each of the year and seasons selected. If the target file
    already exists, the source files are copied only when newer than the target file.
}
\description{
Functions to to read and copy versions of the produksjonstilskuddsregister.
}
\details{
The produksjonstilskuddsregister includes information on number of animals that the produsent has applied subsidies for at the
    counting dates. Since 2017, the counting dates are in March and October. Landbruksdirektoratet provides three to four versions of the
    register for each counting date. The functions automatically selects the last updated version of the register.

    \code{read_Prodtilskudd} Reads the produksjonstilskuddsregister into a data frame. The function gives options to select year and season The standard
    settings will read in the files from NVI's internal network and select the latest updated file for both spring and autumn and combine them
    into one file. If changing the from_path, the function can be used to read the translation file from other directories. This can be useful
    if having a stand alone app with no connection the NVI's internal network. In other cases, it should be avoided.

    \code{copy_Prodtilskudd} copies the source produksjonstilskuddsregister for each of the year and seasons selected to a given directory.
}
\examples{
\dontrun{
# Reading from standard directory at NVI's network
Pkode_last <- read_Prodtilskudd()

# Reading from standard directory at NVI's network and selecting a specific version of the register
Pkode201903 <- read_Prodtilskudd(Pkode_year = "2019", Pkode_month = "03")
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
