% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_PJSdata_for_value-deprecated.R
\name{select_PJSdata_for_value}
\alias{select_PJSdata_for_value}
\title{select_PJSdata_for_value is Deprecated}
\usage{
select_PJSdata_for_value(
  data,
  code_column,
  value_2_check,
  keep_selected = TRUE
)
}
\arguments{
\item{data}{\[`data.frame`\]\cr
PJS data from which a subset should be selected.}

\item{code_column}{\[`character`\]\cr
Vector with the column names for the variables that is used in the selection.}

\item{value_2_check}{\[`character`\]\cr
Vector with the values that should be selected, see details and examples.}

\item{keep_selected}{\[`logical(1)`\]\cr
If `TRUE`, the selected rows are included, if `FALSE`, the selected columns
    are excluded. Defaults to `TRUE`.}
}
\description{
\code{select_PJSdata_for_value} was deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::select_PJSdata_for_value} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
The old help pages can be found at \code{help("select_PJSdata_for_value-deprecated")}.
    Information on deprecated function can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
