% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_disease_parameters-deprecated.R
\name{set_disease_parameters}
\alias{set_disease_parameters}
\title{set_disease_parameters is Deprecated}
\usage{
set_disease_parameters(
  purpose = NULL,
  hensikt2select = NULL,
  hensikt2delete = NULL,
  utbrudd2select = NULL,
  metode2select = NULL,
  analytt2select = NULL,
  analytt2delete = NULL,
  art2select = NULL,
  include_missing_art = NULL,
  FUN = NULL,
  select_statement = NULL,
  selection_parameters = NULL,
  ...
)
}
\arguments{
\item{purpose}{[\code{character}]\cr
A short description of the purpose of the selection, see details. Defaults to NULL.}

\item{hensikt2select}{[\code{character}]\cr
Specific "hensiktkoder" for the "analytt" in question. If sub-codes should
    be included, end the code with \%. Defaults to \code{NULL}.}

\item{hensikt2delete}{[\code{character}]\cr
"hensiktkoder" for which saker should be excluded.
    If sub-codes should be included, end the code with \%. Defaults to \code{NULL}.}

\item{utbrudd2select}{[\code{character(1)}]\cr
"utbruddsID". Defaults to \code{NULL}.}

\item{metode2select}{[\code{character}]\cr
Specific "metodekoder for the "analytt" in question." Defaults to \code{NULL}.}

\item{analytt2select}{[\code{character}]\cr
"analyttkoder" for the agent and/or disease. If sub-codes should be included,
    end the code with \%. Defaults to \code{NULL}.}

\item{analytt2delete}{[\code{character}]\cr
Specific "analyttkoder" that should be deleted, see details. If sub-codes should
    be included, end the code with \%. Defaults to \code{NULL}.}

\item{art2select}{[\code{character}]\cr
"artkoder". If sub-codes should be included, end the code with \%.  \code{NA} can be
    combined with another "artkode". Defaults to \code{NULL}.}

\item{include_missing_art}{[\code{character(1)}]\cr
Should missing art be included. Must be one of c("never", "always", "for_selected_hensikt").
    If NULL, it is set to "always" when \code{art2select} includes NA, else it is set to "never".
    Defaults to \code{NULL}.}

\item{FUN}{[\code{function}]\cr
Function to build the selection statement, see
    \ifelse{html}{\code{\link{retrieve_PJSdata}}}{\code{retrieve_PJSdata}}).
    Defaults to \code{NULL}.}

\item{select_statement}{[\code{character(1)}]\cr
A written select statement, see
    \ifelse{html}{\code{\link{retrieve_PJSdata}}}{\code{retrieve_PJSdata}}).
    Defaults to \code{NULL}.}

\item{selection_parameters}{[\code{character(1)}]\cr
Either the path and file name for an R script that can be sourced and that
    sets the selection parameters or a named list with the selection parameters
    (i.e. equal to the output of this function). Defaults to \code{NULL}.}

\item{\dots}{Other arguments to be passed to `set_disease_parameters`.}
}
\description{
\code{set_disease_parameters} was deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::set_disease_parameters} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
The old help pages can be found at \code{help("set_disease_parameters-deprecated")}.
    Information on deprecated functions can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
