% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_disease_parameters.R
\name{set_disease_parameters}
\alias{set_disease_parameters}
\title{Sets disease selection parameters}
\usage{
set_disease_parameters(
  hensikt2select = NULL,
  hensikt2delete = NULL,
  utbrudd2select = NULL,
  metode2select = NULL,
  analytt2select = NULL,
  art2select = NULL,
  include_missing_art = NULL,
  file = NULL
)
}
\arguments{
\item{hensikt2select}{[\code{character}]\cr
Specific "hensiktkoder" for the "analytt" in question. If sub-codes should
    be included, end the code with \%. Can be \code{NULL}.}

\item{hensikt2delete}{[\code{character}]\cr
"hensiktkoder" for which saker should be excluded
    If sub-codes should be included, end the code with \%. Can be \code{NULL}.}

\item{utbrudd2select}{[\code{character(1)}]\cr
"utbruddsID". Can be \code{NULL}.}

\item{metode2select}{[\code{character}]\cr
Specific "metodekoder for the "analytt" in question." Can be \code{NULL}.}

\item{analytt2select}{[\code{character}]\cr
"analyttkoder". If sub-codes should be included, end the code with \%.
    Can be \code{NULL}.}

\item{art2select}{[\code{character}]\cr
"artkoder". If sub-codes should be included, end the code with \%.  \code{NA} can be
    combined with another "artkode". Can be \code{NULL}.}

\item{include_missing_art}{[\code{character(1)}]\cr
Should missing art be included. Must be one of c("never", "always", "for_selected_hensikt").
    If NULL, it is set to "always" when \code{art2select} includes NA, else it is set to "never".}

\item{file}{[\code{character(1)}]\cr
path and file name for an R script that can be sourced and that
    sets the parameters \code{hensikt2select}, \code{utbrudd2select}, \code{metode2select}, and
    \code{analytt2select}. Can be \code{NULL}.}
}
\value{
A named list with selection parameters that can be used to generate
    SQL selection-statements and facilitate structuring output from PJS.
}
\description{
Sets the disease selection parameters and store them in a list
    object. The list follows a standardised named format and the elements can
    be used as input to 
    \ifelse{html}{\code{\link{build_query_hensikt}}}{\code{build_query_hensikt}},
    \ifelse{html}{\code{\link{build_query_one_disease}}}{\code{build_query_one_disease}}
    or
    \ifelse{html}{\code{\link{build_query_outbreak}}}{\code{build_query_outbreak}}.
}
\details{
Saker in PJS that concern one infection / disease can be characterised
    by the "analytt" (at "konklusjon" and/or "resultat" level), specific "hensikter",
    a relevant "utbrudds_ID" and/or specific "metoder." These can be used to select
    saker in PJS and/or to structure and simplify the output from PJS.

    One or more specific "hensiktkoder" may be input to the selection statement.
    With specific "hensiktkode" is meant a "hensiktkode" that will imply that the sample
    will be examined for specific infectious agent(s) or disease. One or more
    specific "metodekoder" may be input to the selection statement. With specific
    "metodekode" is meant a "metodekode" that implies an examination that will give one
    of the input 2 as a result. If sub-codes of "analyttkode" or "hensiktkode"
    should be included, end the code with \%.

    The selection parameters can be input values for dedicated arguments. For input parameters
    \code{hensikt2select}, \code{utbrudd2select}, \code{metode2select}, 
    \code{analytt2select}, \code{art2select}, and \code{include_missing_art}, 
    the input may be given in a source file. This may be handy if the
    selection will be performed many times. It also gives the possibility of
    using a for loop that selects PJS-data and performs similar analyses for one
    disease at a time.
}
\examples{
# Selection parameters for Pancreatic disease (PD)
selection_parameters <- set_disease_parameters(
  analytt2select = c("01220104\%", "1502010235"),
  hensikt2select = c("0100108018", "0100109003", "0100111003", "0800109"),
  metode2select = c("070070", "070231", "010057", "060265")
  )
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
