% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_PAT-deprecated.R
\name{set_PAT}
\alias{set_PAT}
\alias{get_PAT}
\alias{remove_PAT}
\title{\code{set_PAT}, \code{get_PAT}, and \code{remove_PAT} is deprecated}
\usage{
set_PAT(service)

get_PAT(service)

remove_PAT(service)
}
\arguments{
\item{service}{Name of the internet service, for example "GitHub". For
internet services where one don't use the premade wrappers, the name
can be chosen freely, but must be the same as used in \code{get_PAT}}
}
\description{
\code{set_PAT}, \code{get_PAT}, and \code{remove_PAT} was
    deprecated from NVIdb v0.11.0 released 2023-09-22. These functions
    should be replaced by corresponding functions in package 'gitcreds'
    that are better, more flexible and maintained at cran.
}
\details{
The old help pages can be found at \code{help("set_PAT-deprecated")}.
    Information on deprecated functions can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
