% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_eos_data-deprecated.R
\name{standardize_eos_data}
\alias{standardize_eos_data}
\title{standardize_eos_data is Deprecated}
\usage{
standardize_eos_data(
  data,
  dbsource = deparse(substitute(data)),
  standards = NULL,
  standardize_colnames = TRUE,
  breed_to_species = TRUE,
  adjust_n_examined = TRUE,
  delete_redundant = TRUE,
  ...
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
The data retrieved from EOS.}

\item{dbsource}{[\code{character(1)}]\cr
 If specified, this will be used for fetching standard column names by
\code{\link{standardize_columns}}. Defaults to the name of the input data.}

\item{standards}{[\code{data.frame}]\cr
The translation table to standard column names. Defaults to \code{NULL}.}

\item{standardize_colnames}{[\code{logical(1)}]\cr
If \code{TRUE}, the column names will be standardised. Defaults to \code{TRUE)}.}

\item{breed_to_species}{[\code{logical(1)}]\cr
If \code{TRUE}, breed is translated back to species. Defaults to \code{TRUE)}.}

\item{adjust_n_examined}{[\code{logical(1)}]\cr
If \code{TRUE}, the number of examined samples is adjusted so it is at maximum
the number of received samples. Defaults to \code{TRUE}.}

\item{delete_redundant}{[\code{logical(1)}]\cr
If \code{TRUE}, redundant variables in the data is deleted. Defaults to \code{TRUE}.}

\item{\dots}{Other arguments to be passed to \code{\link{standardize_columns}}.}
}
\description{
\code{standardize_eos_data} was deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::standardize_eos_data} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
The old help pages can be found at \code{help("standardize_eos_data-deprecated")}.
    Information on deprecated function can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
