% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_PJSdata-deprecated.R
\name{standardize_PJSdata-defunct}
\alias{standardize_PJSdata-defunct}
\title{Standardizing PJS-data}
\arguments{
\item{PJSdata}{[\code{data.frame}]\cr
Data retrieved from PJS.}

\item{dbsource}{[\code{character(1)}]\cr
The table that is the source of data. This will be used for fetching
    standard column names by
    \ifelse{html}{\code{\link{standardize_columns}}}{\code{standardize_columns}}
    and should be the name of the data source as registered in the
    "column_standards" table. Defaults to "v2_sak_m_res".}
}
\value{
\code{data.frame} with standardized PJS-data.
}
\description{
Standardizing PJS-data. This standardizing should always be performed.
    Other functions used for further preparation of PJSdata, like
    \ifelse{html}{\code{\link{choose_PJS_levels}}}{\code{choose_PJS_levels}}
    , and
    \ifelse{html}{\code{\link{exclude_from_PJSdata}}}{\code{exclude_from_PJSdata}}
    will not work as intended unless the column names are standardized.
}
\details{
The function performs the following standardizing of data extracted from PJS:
\itemize{
  \item The unnecessary columns konkl_provenr and vet_distriktnr are removed.
  \item The column names are standardized using
    \ifelse{html}{\code{\link{standardize_columns}}}{\code{standardize_columns}}.
  \item Numeric variables are transformed to numbers.
  \item Date variables are transformed to date format.
  \item Character variables are trimmed for leading and trailing spaces.
  \item The variables saksnr and, if possible, fagnr are generated.
  \item Test data, i.e. saker with ansvarlig_seksjon in c("14", "99") are deleted.
  }
}
\author{
Petter Hopp Petter.Hopp@vetinst.no

Johan Åkerstedt Johan.Akerstedt@vetinst.no
}
\keyword{internal}
