% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_PJSdata-deprecated.R
\name{standardize_PJSdata}
\alias{standardize_PJSdata}
\title{standardize_PJSdata is Deprecated}
\usage{
standardize_PJSdata(PJSdata, dbsource = "v2_sak_m_res")
}
\arguments{
\item{PJSdata}{[\code{data.frame}]\cr
Data retrieved from PJS.}

\item{dbsource}{[\code{character(1)}]\cr
The table that is the source of data. This will be used for fetching
    standard column names by
    \ifelse{html}{\code{\link{standardize_columns}}}{\code{standardize_columns}}
    and should be the name of the data source as registered in the
    "column_standards" table. Defaults to "v2_sak_m_res".}
}
\description{
\code{standardize_PJSdata} was deprecated in NVIdb v0.13.0 released
    2024-##-##. All PJS related functions have been moved to \code{NVIpjsr}.
    Use \code{NVIpjsr::standardize_PJSdata} instead. When attaching packages,
    remember to attach \code{NVIdb} before \code{NVIpjsr}.
}
\details{
The old help pages can be found at \code{help("standardize_PJSdata-deprecated")}.
    Information on deprecated function can be found at \code{help("NVIdb-deprecated")}.
}
\keyword{internal}
